### 2025-03-18 #7919 @editorialbot
[PRE REVIEW]: HSI-Wizard: A magical open-source Python package 1 for medical hyperspectral imaging applications
**Submitting author:** @WTFLX (<a href="http://orcid.org/0000-0002-4821-8549">Felix Wühler</a>)
**Repository:** https://github.com/BlueSpacePotato/hsi-wizard
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.12
