### 2025-04-08 89c3d78
added normalization to datacube_ops, added new cluster method 'segment cubes' removed unnessary prints
