+++ wizard/_core/datacube_ops.py
@@ -371,3 +371,24 @@ def remove_vingetting(dc: DataCube, axis: int = 1, slice_params: dict = None) ->
     dc.set_cube(corrected_cube)
 
     return dc
+
+
+def normalize(dc) -> DataCube:
+        """
+    Normalizing the the spectral information between 0 and 1
+
+    Parameters
+    ----------
+    dc : DataCube
+    The DataCube instance to process.
+
+    Returns
+    -------
+    DataCube
+    The processed DataCube.
+        """
+        cube = dc.cube - dc.cube.min()
+        cube = cube / cube.max()
+        dc.set_cube(cube)
+
+        return dc
