### 2025-04-29 #7919 @Kevin-Mattheus-Moerman
@WTFLX - thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editors most appropriate for your area are all already rather busy.
