+++ wizard/_utils/helper.py
@@ -20,7 +20,16 @@ Functions
 """
 
 import cv2
+import warnings
 import numpy as np
+from skimage.feature import canny
+
+
+
+class RegistrationError(Exception):
+    """Custom exception for registration errors."""
+
+    pass
 
 
 def find_nex_greater_wave(waves, wave_1: int, maximum_deviation: int = 5) -> int:
@@ -81,7 +90,7 @@ def find_nex_smaller_wave(waves, wave_1: int, maximum_deviation: int = 5) -> int
     return wave_next
 
 
-def normalize(spec):
+def normalize_spec(spec):
     """Normalize the spectrum to the range 0-1 if needed."""
     spec_min, spec_max = spec.min(), spec.max()
     return np.clip((spec - spec_min) / (spec_max - spec_min), 0, 1) if spec_max > spec_min else spec
@@ -140,3 +149,167 @@ def feature_regestration(o_img: np.ndarray, a_img: np.ndarray, max_features: int
     aligned_img = cv2.warpPerspective(a_img, h, (width, height))
 
     return aligned_img, h
+
+def feature_registration(o_img: np.ndarray, a_img: np.ndarray,
+                          max_features: int = 5000,
+                          match_percent: float = 0.1
+                          ) -> tuple[np.ndarray, np.ndarray]:
+    """
+    Perform ORB-based feature registration.
+
+    Aligns `a_img` to `o_img` using ORB features and RANSAC
+    to find the homography matrix.
+
+    Parameters
+    ----------
+    o_img : numpy.ndarray
+        The target (reference) image.
+    a_img : numpy.ndarray
+        The source image to be aligned.
+    max_features : int, optional
+        Maximum number of ORB features to detect, defaults to 5000.
+    match_percent : float, optional
+        Percentage of best matches to use for homography, defaults to 0.1.
+
+    Returns
+    -------
+    aligned_img : numpy.ndarray
+        The `a_img` warped to align with `o_img`.
+    H : numpy.ndarray
+        The 3x3 homography matrix mapping points from `a_img` to `o_img`.
+
+    Raises
+    ------
+    RegistrationError
+        If no descriptors are found, not enough good matches are found,
+        or homography estimation fails.
+    """
+    def to_uint8(im: np.ndarray) -> np.ndarray:
+        """Convert image to uint8, normalizing if necessary."""
+        if im.dtype != np.uint8:
+            with warnings.catch_warnings():
+                warnings.filterwarnings("ignore", category=RuntimeWarning, message="invalid value encountered in true_divide")
+                warnings.filterwarnings("ignore", category=RuntimeWarning, message="invalid value encountered in divide")
+                min_val, max_val = im.min(), im.max()
+                if min_val == max_val:
+                    im_norm = np.zeros_like(im, dtype=np.float32)
+                else:
+                    im_norm = (im.astype(np.float32) - min_val) / (max_val - min_val)
+                im_norm = np.nan_to_num(im_norm, nan=0.0, posinf=1.0, neginf=0.0)
+            im = (im_norm * 255).astype(np.uint8)
+        return im
+
+    o8, a8 = to_uint8(o_img), to_uint8(a_img)
+    orb = cv2.ORB_create(max_features)
+    kp1, des1 = orb.detectAndCompute(a8, None)
+    kp2, des2 = orb.detectAndCompute(o8, None)
+
+    if des1 is None or des2 is None:
+        raise RegistrationError("No descriptors found")
+
[…]

[…]
