### 2025-05-14 45a0c5d
clean up styling and move _process_slices to testhelper
