+++ wizard/_utils/tracker.py
@@ -18,28 +18,38 @@ Classes
    :members:
    :undoc-members:
    :show-inheritance:
-
 """
 
-exculted = ['stop_recording', 'save_template', '_clean_data', '_map_args_to_kwargs', 'execute_template']
+excluded = ['stop_recording', 'save_template', '_clean_data', '_map_args_to_kwargs', 'execute_template']
 
 
 class TrackExecutionMeta(type):
     """
-    A metaclass for tracking method executions in classes that use it (e.g., DataCube).
+    Metaclass for tracking method executions in classes that use it.
 
-    This metaclass automatically wraps methods to log their calls when recording is enabled.
-    Only methods marked as dynamic (with `__is_dynamic__ = True`) are recorded.
-    Useful for debugging, auditing, or dynamically analyzing method usage patterns.
+    Automatically wraps dynamic methods (marked with `__is_dynamic__ = True`)
+    to log their calls when tracking is enabled. Used for debugging, auditing,
+    or analyzing usage patterns of dynamic methods.
 
-    Attributes:
-        recording (bool): Flag indicating whether to track method calls.
-        recorded_methods (list): Stores tuples of (method_name, args, kwargs) for recorded calls.
+    Attributes
+    ----------
+    recording : bool
+        Indicates whether method tracking is currently active.
 
-    Methods:
-        start_recording(): Enables method tracking and resets previous records.
-        stop_recording(): Disables method tracking.
-        record_method(func): Decorator to wrap and conditionally log method calls.
+    recorded_methods : list of tuple
+        Stores a list of tuples (method_name, args, kwargs) representing
+        recorded method calls.
+
+    Methods
+    -------
+    start_recording()
+        Enables method tracking and clears previously recorded data.
+
+    stop_recording()
+        Disables method tracking.
+
+    record_method(func)
+        Decorator to wrap and conditionally record dynamic method calls.
     """
 
     recording = False
@@ -47,18 +57,27 @@ class TrackExecutionMeta(type):
 
     def __new__(cls, name, bases, dct):
         """
-        Overrides the class creation process to wrap methods for tracking.
+        Creates a new class, wrapping its methods for tracking if applicable.
 
-        Args:
-            name (str): Name of the class being created.
-            bases (tuple): Base classes of the new class.
-            dct (dict): Dictionary of attributes/methods in the class.
+        Wraps all callable attributes except those explicitly excluded or not marked dynamic.
 
-        Returns:
-            type: A new class with wrapped methods for execution tracking.
+        Parameters
+        ----------
+        name : str
+            Name of the class being created.
+
+        bases : tuple
+            Base classes of the new class.
+
+        dct : dict
+            Dictionary of the class's attributes and methods.
+
+        Returns
+        -------
+        type
+            A new class with method execution tracking wrappers.
         """
         for key, value in dct.items():
-            # Wrap only dynamic methods or those that are not in the excluded list
             if callable(value) and key != 'execute_template':
                 dct[key] = cls.record_method(value)
         return super().__new__(cls, name, bases, dct)
@@ -66,13 +85,24 @@ class TrackExecutionMeta(type):
     @staticmethod
     def record_method(func):
         """
-        Decorator to wrap a method and record its execution if it's marked as dynamic.
+        Wraps a method to record its execution if tracking is enabled.
 
-        Args:
-            func (callable): The function to be wrapped.
+        Only records calls to methods explicitly marked as dynamic
+        using the `__is_dynamic__ = True` attribute.
 
-        Returns:
-            callable: A wrapped version of the original function.
+        Parameters
+        ----------
+        func : callable
+            The method to be wrapped.
+
+        Returns
+        -------
+        callable
[…]
