### 2025-05-15 #7919 @csoneson
👋🏻 @ayoubft, @tomelse, @mh-skjelvareid - would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> HSI-Wizard: A magical open-source Python package for medical hyperspectral imaging applications
#7919

You can find more information about our checklist-driven review process [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

Thanks in advance!
