### 2025-05-16 0b9226f
fixed remove_spikes, added dc.registered var, update merge to check dc.registered, added image loader, fixed error with folder and image loader for 1 channel images
