+++ wizard/_core/datacube_ops.py
@@ -15,11 +15,13 @@ This module contains operation function for processing datacubes.
 import cv2
 import copy
 import rembg
+import random
 import numpy as np
 from PIL import Image
 from joblib import Parallel, delayed
 from scipy.signal import savgol_filter
 from scipy.ndimage import gaussian_filter
+from skimage.transform import warp
 
 from . import DataCube 
 from .._processing.spectral import calculate_modified_z_score, spec_baseline_als
@@ -28,35 +30,69 @@ from .._utils.helper import _process_slice, feature_registration, RegistrationEr
 
 def remove_spikes(dc: DataCube, threshold: int = 6500, window: int = 5) -> DataCube:
     """
-    Remove cosmic spikes from the DataCube.
+    Remove cosmic spikes from each pixel's spectral data.
 
-    Identifies spikes using the modified z-score and replaces them
-    with the mean of neighboring values within the specified window.
+    This function computes the modified z-score for each pixel's spectral vector,
+    identifies spikes where the score exceeds the threshold, and replaces spike
+    values with the local mean within a sliding window along the spectrum.
 
     Parameters
     ----------
     dc : DataCube
-        The input DataCube from which cosmic spikes are to be removed.
+        The input DataCube with shape (v, x, y), where v is the number of spectral bands.
     threshold : int, optional
         Threshold for spike detection via modified z-score, defaults to 6500.
     window : int, optional
-        Window size for mean calculation when replacing spikes, defaults to 5.
+        Window size (in spectral channels) for mean replacement of spikes, defaults to 5.
 
     Returns
     -------
     DataCube
-        The DataCube with spikes removed.
+        A new DataCube instance with spikes removed per-pixel.
+
+    Raises
+    ------
+    ValueError
+        If `window` is not in the range [1, number of spectral bands].
+
+    Notes
+    -----
+    - The original DataCube is not modified in place; es wird eine Kopie zurückgegeben.
+    - Die Modifizierte z-Score-Berechnung erwartet Input mit Form (n_samples, n_features).
+    - Parallelisierung beschleunigt die Einzelpixel-Bearbeitung.
+
+    Examples
+    --------
+    >>> dc = DataCube.read("example.fsm")
+    >>> dc_clean = remove_spikes(dc, threshold=6500, window=5)
     """
-    z_spectrum = calculate_modified_z_score(dc.cube.reshape(dc.shape[0], -1))
-    spikes = abs(z_spectrum) > threshold
-    cube_out = dc.cube.copy()
-    spec_out_flat = cube_out.reshape(cube_out.shape[0], -1)
+    v, x, y = dc.cube.shape
+    if not (1 <= window <= v):
+        raise ValueError(f"window must be between 1 and {v}, got {window}")
+
+    # reshape to (n_pixels, v)
+    n_pixels = x * y
+    flat_cube = dc.cube.reshape(v, n_pixels).T  # shape: (n_pixels, v)
+
+    # Berechne pro-pixel modifizierten z-score
+    z_scores = calculate_modified_z_score(flat_cube)  # (n_pixels, v)
+    spikes = np.abs(z_scores) > threshold
+    flat_out = flat_cube.copy()
+
+    # Parallel auf jedes Pixel anwenden
     results = Parallel(n_jobs=-1)(
-        delayed(_process_slice)(spec_out_flat, spikes, idx, window)
-        for idx in range(spikes.shape[0]))
-    for idx, tmp in results:
-        spec_out_flat[idx] = tmp
-    dc.set_cube(spec_out_flat.reshape(dc.shape))
+        delayed(_process_slice)(flat_out, spikes, idx, window)
[…]
