### 2025-05-16 #8207 @editorialbot
[REVIEW]: HSI-Wizard: A magical open-source Python package 1 for medical hyperspectral imaging applications
**Submitting author:** @BlueSpacePotato (<a href="http://orcid.org/0000-0002-4821-8549">Felix Wühler</a>)
**Repository:** https://github.com/BlueSpacePotato/hsi-wizard
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.42
