### 2025-05-17 62f5083
added dc.registered to regoster method
