### 2025-05-19 bcaa995
update testcases and styling, added support to plot all surfaces in the faces.py script
