+++ tests/test_exploration.py
@@ -1,30 +1,166 @@
-import random
-
 import numpy as np
 import matplotlib
+import pytest
+import matplotlib.pyplot as plt
+from mpl_toolkits import mplot3d
 
 import wizard
+from wizard import plotter
 from wizard._exploration import plotter
-
-import pytest
+from wizard import plot_datacube_faces
+from wizard import DataCube
 
 matplotlib.use("Agg")
 
-class TestPlotter:
 
-    def test_plot_function_runs(self):
-        dc = wizard.DataCube(cube=np.random.rand(20, 8, 9), wavelengths=np.random.randint(0, 200, size=20))
-        try:
-            wizard.plotter(dc)
-            assert True
-        except:
-            assert False
+def create_test_cube(shape=(3, 4, 4)):
+    cube = np.random.rand(*shape)  # Generate a random cube
+    return DataCube(cube=cube, wavelengths=np.arange(shape[0]))
 
-    def test_normalize_small_range(self):
-        spec = np.array([0.0001, 0.0002, 0.0003])
-        normalized = plotter.normalize_spec(spec)
-        expected = np.array([0.0, 0.5, 1.0])
-        np.testing.assert_allclose(normalized, expected, atol=1e-6)
+
+def setup_plot(dc):
+    # reset global state and any saved lists
+    plotter.state['layer_id'] = 0
+    plotter.state['normalize_flag'] = False
+    plotter.saved_plots.clear()
+    plotter.saved_lines.clear()
+    plotter.saved_rois.clear()
+
+    # launch headless plotter
+    wizard.plotter(dc)
+    return plt.gcf()
+
+
+def get_key_callbacks(fig):
+    """
+    Return only the user-registered on_key callbacks.
+    """
+    refs = fig.canvas.callbacks.callbacks.get('key_press_event', {}).values()
+    cbs = []
+    for ref in refs:
+        func = ref()
+        if func and func.__name__ == 'on_key':
+            cbs.append(func)
+    if not cbs:
+        raise RuntimeError("on_key callback not found")
+    return cbs
+
+
+def get_button_press_callbacks(fig):
+    refs = fig.canvas.callbacks.callbacks.get('button_press_event', {}).values()
+    cbs = []
+    for ref in refs:
+        func = ref()
+        if func is not None:
+            cbs.append(func)
+    return cbs
+
+
+def get_onclick_select_callback(fig):
+    for cb in get_button_press_callbacks(fig):
+        if cb.__name__ == 'onclick_select':
+            return cb
+    raise RuntimeError("onclick_select callback not found")
+
+
+class TestPlotterExtra:
+
+    def test_on_key_navigation_bounds(self):
+        dc = create_test_cube((3, 2, 2))
+        fig = setup_plot(dc)
+        key_cbs = get_key_callbacks(fig)
+
+        # initial state
+        assert plotter.state['layer_id'] == 0
+
+        # press right twice: should move to last index
+        evt = type("E", (), {"key": "right"})
+        for _ in range(2):
+            for cb in key_cbs:
+                cb(evt)
+        assert plotter.state['layer_id'] == 2
+
+        # pressing right again stays at upper bound
+        for cb in key_cbs:
+            cb(evt)
+        assert plotter.state['layer_id'] == 2
+
+        # press left thrice: should move back to 0
+        evt.key = "left"
+        for _ in range(3):
+            for cb in key_cbs:
+                cb(evt)
+        assert plotter.state['layer_id'] == 0
+
+    def test_update_plot_title_and_vertical_line(self):
+        cube = np.random.rand(4, 2, 2)
[…]
