+++ tests/test_core.py
@@ -12,6 +12,7 @@ def create_test_cube(shape=(3, 4, 4)):
     cube = np.random.rand(*shape)  # Generate a random cube
     return DataCube(cube=cube, wavelengths=np.arange(shape[0]))
 
+
 class TestDataCube:
 
     #  DataCube can be initialized with cube, wavelengths, name, and notation.
@@ -572,7 +573,7 @@ class TestDataCubeOps:
         dc = create_test_cube(shape=(3, 24, 24))
 
         # Create a temporary cube with specific values for testing
-        tmp_cube = dc.cube.copy() * 0.6
+        tmp_cube = dc.cube.copy() * 0.2
 
         # Define the middle region that should remain unchanged
         x_start, x_end = int(dc.shape[1] * 0.25), int(dc.shape[1] * 0.75)
@@ -580,8 +581,8 @@ class TestDataCubeOps:
 
         # Apply specific values to the middle region using broadcasting
         tmp_cube[0, x_start:x_end, y_start:y_end] = 0.8
-        tmp_cube[1, x_start:x_end, y_start:y_end] = 0.7
-        tmp_cube[2, x_start:x_end, y_start:y_end] = 0.9
+        tmp_cube[1, x_start:x_end, y_start:y_end] = 0.8
+        tmp_cube[2, x_start:x_end, y_start:y_end] = 0.8
 
         # Set the modified cube
         dc.set_cube(tmp_cube)
@@ -680,3 +681,100 @@ class TestDataCubeOps:
         with pytest.raises(NotImplementedError):
             dc1.merge_cubes(dc2)
 
+    def test_register_layers_simple(self):
+        import numpy as np
+        dc = create_test_cube(shape=(3, 5, 5))
+        # Make all layers identical for easy registration
+        base = np.random.rand(5, 5)
+        dc.cube = np.stack([base, base, base])
+        result = dc.register_layers_simple(max_features=1000, match_percent=0.2)
+        assert result is dc
+        assert dc.cube.shape == (3, 5, 5)
+        assert getattr(dc, 'registered', False) is True
+
+    def test_remove_vignetting_poly_invalid_axis(self):
+        dc = create_test_cube()
+        with pytest.raises(ValueError):
+            dc.remove_vignetting_poly(axis=3)
+
+    def test_remove_vignetting_poly_default(self):
+        import numpy as np
+        dc = create_test_cube(shape=(2, 72, 72))
+        original_shape = dc.cube.shape
+        result = dc.remove_vignetting_poly()
+        assert result is dc
+        assert dc.cube.shape == original_shape
+        # After polynomial vignetting removal, dtype should be float32
+        assert dc.cube.dtype == np.float32
+
+    def test_normalize(self):
+        import numpy as np
+        dc = create_test_cube(shape=(2, 4, 4))
+        data = np.array([
+            [[ 1,  2,  3,  4],
+             [ 5,  6,  7,  8],
+             [ 9, 10, 11, 12],
+             [13, 14, 15, 16]],
+            [[16, 15, 14, 13],
+             [12, 11, 10,  9],
+             [ 8,  7,  6,  5],
+             [ 4,  3,  2,  1]]
+        ], dtype=float)
+        dc.set_cube(data)
+        dc.normalize()
+        # Each layer should now span [0,1]
+        for layer in dc.cube:
+            assert np.isclose(layer.min(), 0.0)
+            assert np.isclose(layer.max(), 1.0)
+
+    def test_normalize_polarity_uint8(self):
+        import numpy as np
+        from wizard._core.datacube_ops import normalize_polarity
+        img = np.full((10, 10), 200, dtype=np.uint8)
+        out = normalize_polarity(img)
+        assert out.dtype == np.float32
[…]
