### 2025-05-20 8e86332
adjustet tests for more reliability outputs
