### 2025-05-20 08cf596
moved normalize polarity to _utils.helper, added more examples in docstrings
