### 2025-05-20 8d8e8c7
added more examples
