### 2025-05-20 7dfb85b
update helper.py docstrings
