+++ wizard/_utils/helper.py
@@ -33,21 +33,42 @@ class RegistrationError(Exception):
 
 def find_nex_greater_wave(waves, wave_1: int, maximum_deviation: int = 5) -> int:
     """
-    Finds the next greater wave value in a list of waves within a specified deviation.
+    Find the next greater wave value within a specified deviation.
 
-    This function identifies the smallest wave value greater than the specified `wave_1`
-    within a range defined by `maximum_deviation`. If no such value exists, it returns -1.
+    This function searches for the smallest wave value greater than or equal to `wave_1`
+    that exists in the `waves` list, within a deviation range up to `maximum_deviation`.
+    If no such wave is found, the function returns -1.
 
-    :param waves: A list of integers representing the available wave values.
-    :type waves: list[int]
-    :param wave_1: The starting wave value to find the next greater wave for.
-    :type wave_1: int
-    :param maximum_deviation: The maximum deviation from `wave_1` to consider.
-    :type maximum_deviation: int
-    :returns: The next greater wave value within the deviation range, or -1 if no such value exists.
-    :rtype: int
+    Parameters
+    ----------
+    waves : list[int]
+        A list of integer wave values to search within.
+    wave_1 : int
+        The reference wave value to find the next greater wave after.
+    maximum_deviation : int, optional
+        The maximum positive offset from `wave_1` to consider (default is 5).
+
+    Returns
+    -------
+    int
+        The next greater wave value found within the range, or -1 if none exists.
+
+    Raises
+    ------
+    None
+
+    Notes
+    -----
+    The function stops searching once it finds the first match within the allowed range.
+
+    Examples
+    --------
+    >>> find_nex_greater_wave([400, 405, 410, 415], 403)
+    405
+
+    >>> find_nex_greater_wave([400, 405, 410, 415], 416)
+    -1
     """
-
     wave_next = -1
 
     for n in range(maximum_deviation):
@@ -60,23 +81,45 @@ def find_nex_greater_wave(waves, wave_1: int, maximum_deviation: int = 5) -> int
     return wave_next
 
 
+
 def find_nex_smaller_wave(waves, wave_1: int, maximum_deviation: int = 5) -> int:
     """
-    Finds the next smaller wave value in a list of waves within a specified deviation.
+    Find the next smaller wave value within a specified deviation.
 
-    This function identifies the largest wave value smaller than the specified `wave_1`
-    within a range defined by `maximum_deviation`. If no such value exists, it returns -1.
+    This function searches for the largest wave value smaller than or equal to `wave_1`
+    that exists in the `waves` list, within a deviation range up to `maximum_deviation`.
+    If no such wave is found, the function returns -1.
 
-    :param waves: A list of integers representing the available wave values.
-    :type waves: list[int]
-    :param wave_1: The starting wave value to find the next smaller wave for.
-    :type wave_1: int
-    :param maximum_deviation: The maximum deviation from `wave_1` to consider.
-    :type maximum_deviation: int
-    :returns: The next smaller wave value within the deviation range, or -1 if no such value exists.
-    :rtype: int
+    Parameters
+    ----------
+    waves : list[int]
+        A list of integer wave values to search within.
+    wave_1 : int
+        The reference wave value to find the next smaller wave before.
+    maximum_deviation : int, optional
+        The maximum negative offset from `wave_1` to consider (default is 5).
+
+    Returns
+    -------
+    int
+        The next smaller wave value found within the range, or -1 if none exists.
+
+    Raises
+    ------
+    None
+
+    Notes
+    -----
[…]
