### 2025-05-21 96a3b7a
added more upscale methods
