+++ wizard/_core/datacube_ops.py
@@ -12,6 +12,7 @@ This module contains operation function for processing datacubes.
 .. autofunction:: remove_spikes
 .. autofunction:: resize
 """
+import os
 import cv2
 import copy
 import rembg
@@ -22,6 +23,7 @@ from joblib import Parallel, delayed
 from scipy.signal import savgol_filter
 from scipy.ndimage import gaussian_filter
 from skimage.transform import warp
+
 
 from . import DataCube 
 from .._processing.spectral import calculate_modified_z_score, spec_baseline_als
@@ -557,9 +559,6 @@ def normalize(dc: DataCube) -> DataCube:
     return dc
 
 
-
-
-
 def register_layers_best(
         dc: DataCube,
         ref_layer: int = 0,
@@ -736,4 +735,399 @@ def remove_vignetting(dc: DataCube, sigma: float = 50, clip: bool = True, epsilo
             corrected_band = np.clip(corrected_band, info.min, info.max)
         corrected_cube[i] = corrected_band.astype(orig_dtype)
     dc.set_cube(corrected_cube)
+    return dc
+
+
+def upscale_datacube_edsr(dc: DataCube, scale: int, model_path: str):
+    """
+    Upscale the spatial dimensions of a DataCube using the EDSR super-resolution model.
+
+    This function applies the EDSR (Enhanced Deep Super-Resolution) model from OpenCV’s
+    dnn_superres module to each spectral slice of the given DataCube. Since EDSR
+    expects a 3-channel input, each single-band slice is duplicated across three
+    channels before processing. The output slices are then reassembled into a new
+    DataCube with upscaled spatial dimensions.
+
+    Parameters
+    ----------
+    dc : DataCube
+        An instance of the DataCube class. Must have attributes:
+        - datacube.cube: numpy array of shape (v, x, y)
+        - datacube.wavelength: list of length v
+    scale : int
+        The upscaling factor (e.g., 2, 3, or 4) supported by the EDSR model.
+    model_path : str
+        Filesystem path to the pretrained EDSR `.pb` model file
+        (e.g., `"EDSR_x4.pb"`).
+
+    Returns
+    -------
+    DataCube
+        A new DataCube instance whose `.cube` has shape
+        (v, x * scale, y * scale) and the same `.wavelength` list.
+
+    Raises
+    ------
+    FileNotFoundError
+        If the specified `model_path` does not exist or is unreadable.
+    cv2.error
+        If OpenCV fails to load the model or perform upsampling.
+    ValueError
+        If the `scale` is not one of the factors supported by the loaded model.
+
+    Notes
+    -----
+    - Requires `opencv-contrib-python>=4.3.0`.
+    - Processing is done slice-by-slice and may be slow for large cubes.
+    - Memory usage grows by approximately `scale^2`.
+
+    Examples
+    --------
+    >>> import wizard
+    >>> dc = wizard.read("hyperspectral.fsm")
+    >>> dc.upscale_datacube_edsr(scale=4, model_path="EDSR_x4.pb")
+    >>> print(dc.cube.shape)
+    (v, x*4, y*4)
+    """
+    # Create EDSR super-res object
+    if not os.path.isfile(model_path):
+        raise FileNotFoundError(f"Model file not found: {model_path}")
+    sr = cv2.dnn_superres.DnnSuperResImpl_create()
+    sr.readModel(model_path)
+    sr.setModel("edsr", scale)
+
+    v, x, y = dc.shape
+    new_x, new_y = x * scale, y * scale
+    up_cube = np.empty((v, new_x, new_y), dtype=dc.cube.dtype)
+
+    for i in range(v):
+        # Extract single-band slice
+        band = dc.cube[i, :, :]
[…]
