### 2025-05-23 BlueSpacePotato/hsi-wizard#14 @mh-skjelvareid
Add use of real-word hyperspectral data to examples
This issue is part of the review the paper submitted to the Journal of Open Source Software: https://github.com/openjournals/joss-reviews/issues/8207

All examples are based on random data, which can be generated easily, but don't represent real data very well. Doing data manipulation like "spike removal" and image resizing on random numbers doesn't give the reader any intuition for why these are needed. It would have been nice, especially for the plotting examples, to show some kind of real-world data.
