### 2025-05-23 BlueSpacePotato/hsi-wizard#15 @mh-skjelvareid
Add explanation and code output to examples
This issue is part of the review of the paper submitted to the Journal of Open Source Software: https://github.com/openjournals/joss-reviews/issues/8207

The examples in the documentation give a quick introduction to how HSI-Wizard is used, and also gives the reader a good intuition for how it works. However, the examples only consist of code with a few comments. The examples would benefit from some further explanation (text outside code block) of what the code does. Also, examples including plots should show the generated plots, and examples including text output should show the text output.
