### 2025-05-23 BlueSpacePotato/hsi-wizard#16 @mh-skjelvareid
Expand community guidelines
This issue is part of the review of the paper submitted to the Journal of Open Source Software: https://github.com/openjournals/joss-reviews/issues/8207

The README file includes information about how to contribute via pull requests. The description is two sentences long and seems a bit "minimal". There are no guidelines on reporting issues or seeking support. I suggest expanding the "contributing" section and adding a section (or two) on reporting issues and seeking support.
