### 2025-05-23 BlueSpacePotato/hsi-wizard#17 @mh-skjelvareid
Add wavelength selector in plotter
This issue is part of the review of the paper submitted to the Journal of Open Source Software: https://github.com/openjournals/joss-reviews/issues/8207

The surface plotter has a slider to select which wavelength is displayed. The plotter seems like it should have something similar, but there is no slider in the user interface. I suggest adding such a slider (or adding user instructions if there are other ways to change wavelength).

![Image](https://github.com/user-attachments/assets/b74a3052-776e-42d1-964a-1883fa1cd2d6)
