### 2025-05-23 BlueSpacePotato/hsi-wizard#18 @mh-skjelvareid
Add examples showcasing the features of spectral and cluster modules
This issue is part of the review of the paper submitted to the Journal of Open Source Software: https://github.com/openjournals/joss-reviews/issues/8207

The "spectral" and "cluster" modules are only described in terms of function docstrings. Adding examples to the documentation demonstrating the features of these modules would showcase both how they are used and how they are useful.
