### 2025-05-23 BlueSpacePotato/hsi-wizard#19 @mh-skjelvareid
Define intended audience more clearly in README / documentation
This issue is part of the review of the paper submitted to the Journal of Open Source Software: https://github.com/openjournals/joss-reviews/issues/8207

The README.md and documentation gives a description of the package and its features, but it feels very brief. It is clear that the package is used for working with hyperspectral data, but what kind of hyperspectral data? The JOSS paper title includes the term "medical hyperspectral imaging applications", but this is not mentioned in the documentation. Note that the largest user group for hyperspectral images is probably within the field of remote sensing (where handling of geospatial coordinates is important). The package should describe its intended audience more clearly.

As in the paper, the documentation should also briefly compare the software to existing solutions, and describe the "niche" that it fills.
