### 2025-05-23 BlueSpacePotato/hsi-wizard#20 @mh-skjelvareid
Clean up "create template" example
This issue is part of the review of the paper submitted to the Journal of Open Source Software: https://github.com/openjournals/joss-reviews/issues/8207

The "create template" example includes wizard.plotter(dc) on lines 9 and 16. When I run the code, a plotting window is opened, and when I close it, the terminal reports "RuntimeError: main thread is not in main loop". A second plot is opened, and I can see that the image has been resized, but using the interactive window to show this is a bit clunky. Also, when I close the window, another error pops up: "  File "C:\Users\mha114\Miniconda3\envs\hsi-wizard\Lib\site-packages\joblib\externals\loky\backend\resource_tracker.py", line 181, in __del__, AttributeError: 'super' object has no attribute '__del__'".

If the point is to demonstrate the use of templates, I would drop plotting the data and focus on recording and saving of the template.
