### 2025-05-23 BlueSpacePotato/hsi-wizard#21 @mh-skjelvareid
Explain intended usage of "surface plotter" in documentation
This issue is part of the review of the paper submitted to the Journal of Open Source Software: https://github.com/openjournals/joss-reviews/issues/8207

It is unclear (at least to me) what the intended usage of the "surface plotter" tool is. Is it a way to find thresholds for masking images? Why, then, is the data visualized as a surface and not as an image? I suggest expanding the documentation of this feature. Using real-world data in an example would also be useful.
