### 2025-05-23 #8207 @mh-skjelvareid
# Comments
First of all I'd like to say that HSI-Wizard is very solid work, following best practices for Python packages, including good test coverage, easy installation, good API documentation and instructive examples. Impressive! 👏 I've listed all my comments below, structured by the checklist items that I haven't checked yet. I have also created [issues in the hsi-wizard repository](https://github.com/BlueSpacePotato/hsi-wizard/issues) related to software and documentation. Comments on the paper are only listed below.

## Functionality
The README file states that the HSI-Wizard "supports everything from basic spectral analysis to advanced machine learning and AI methods." It is unclear what is meant by machine learning and AI in this context. The package supports clustering (unsupervised learning) and manipulation of spectral data, but "machine learning" and "AI" does not seem appropriate here. Except if the word "supports" is meant as enabling subsequent machine learning and AI (outside the package) by pre-processing hyperspectral data?

The paper also emphasizes that HSI-Wizard is meant as tool to unify a diverse set of hyperspectral datasets. If this is the case, I think the list of supported file formats should be extended to include some more of the most common file formats for hyperspectral data. My background is not in medical imaging, so I don't know the most common file formats in this field, but I that think at least support for [ENVI files](https://www.nv5geospatialsoftware.com/docs/enviheaderfiles.html) would be a natural addition (which is relatively easy via [Spectral Python](http://www.spectralpython.net/fileio.html#envi-headers)). I don't see this as an acceptance-blocker, as I think the authors are free to set the scope of their own software, but it would significantly expand the audience for the package.

I've added issues for some other functionality details:
- https://github.com/BlueSpacePotato/hsi-wizard/issues/17
- https://github.com/BlueSpacePotato/hsi-wizard/issues/20
- https://github.com/BlueSpacePotato/hsi-wizard/issues/21

## Documentation

### A statement of need
The README.md and documentation gives a description of the package and its features, but it feels very brief. It is clear that the package is used for working with hyperspectral data, but what kind of hyperspectral data? The JOSS paper title includes the term "medical hyperspectral imaging applications", but this is not mentioned in the documentation. Note that the largest user group for hyperspectral images is probably within the field of remote sensing (where handling of geospatial coordinates is important). The package should describe its intended audience more clearly.

As in the paper, the documentation should also briefly compare the software to existing solutions (see "state of the field" below), and describe the "niche" that it fills.

- https://github.com/BlueSpacePotato/hsi-wizard/issues/19

### Installation instructions:
The installation of hsi-wizard via pip is straightforward and follow good Python practices. The checklist asks "Is there a clearly-stated list of dependencies?". The package depends on e.g. NumPy, SciPy, scikit-image, scikit-learn, OpenCV,  etc. - these are listed in pyproject.toml file, but are not mentioned in the documentation. @csoneson , I'm unsure - what are the requirements here? The total size of installed packages is almost 1 GB, this should also be mentioned in the documentation.

### Example usage
The examples give a quick introduction to how HSI-Wizard is used, and also gives the reader a good intuition for how it works. However, the examples only consist of code with a few comments. The examples would benefit from some further explanation (text outside code block) of what the code does. Also, examples including plots should show the generated plots, and examples including text output should show the text output.
- https://github.com/BlueSpacePotato/hsi-wizard/issues/15

All examples are based on random data, which can be generated easily, but don't represent real data very well. Doing data manipulation like "spike removal" and image resizing on random numbers doesn't give the reader any intuition for why these are needed. It would have been nice, especially for the plotting examples, to show some kind of real-world data.
- https://github.com/BlueSpacePotato/hsi-wizard/issues/14

The "spectral" and "cluster" modules are only described in terms of the function docstrings. Adding examples demonstrating the features of these modules would showcase both how they are used and how they are useful.
- https://github.com/BlueSpacePotato/hsi-wizard/issues/18

### Community guidelines
The README files includes information about how to contribute via pull requests. The description is two sentences long and seems a bit "minimal". There are no guidelines on reporting issues or seeking support.
- https://github.com/BlueSpacePotato/hsi-wizard/issues/16

## Software paper
### A statement of need
The paper describes data fusion and logging of data manipulations as two of the most important features of the software. I think this description can be expanded to give the readers a better understanding of why this is needed, and how the `DataCube` class helps solve these problems. A short example using real-world data would be instructive (in the paper and/or in the README file).

### State of the field
The [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) states that the paper should include "A description of how this software compares to other commonly-used packages in this research area.". For some reason, this point is not included in the [author guidelines](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) (?). However, the guidelines say that the statement of need should place the software "in the context of related work". I think the paper would benefit from a comparison with similar software, such as

- [Spectral Python](https://www.spectralpython.net/)
- [Spectronon](https://resonon.com/blog-spectronon-hyperspectral-imaging-software)
- [IDCubeLite](https://www.idcubes.com/idcubelite)
- [GLIMPS](https://www.rese-apps.com/software/glimps/index.html)

Note that most of these are not open source - the authors can take the opportunity to argue the advantages of HSI-Wizard being open source.

### References
Some of the references in [paper.bib](https://github.com/BlueSpacePotato/hsi-wizard/blob/joss/joss/paper.bib) are not validly formatted (see editorialbot output above). The [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) also states that the paper should include "A list of key references *including a link to the software archive.*". @csoneson  I'm not sure - should the references include a link to the GitHub repo?
