### 2025-05-23 a971b74
docs: update plotter documentation and docstring; added an Interactive Controls section (closes #17)
