+++ docs/source/modules/exploration/plotter.rst
@@ -2,45 +2,47 @@
 
 Plotter
 -------
-.. module::  wizard._exploration.plotter
+
+.. module:: wizard._exploration.plotter
    :platform: Unix
-   :synopsis: Plotter function for hyperspectral imaging (HSI) data.
+   :synopsis: Interactive plotting interface for hyperspectral DataCube objects.
 
-.. function:: plotter(dc)
+Module Overview
+===============
 
-   This module provides an interactive plotting interface to explore and analyze
-   data cubes. Users can visualize slices, define regions of interest (ROIs), and
-   inspect the spectral data interactively. It features saving, removing, and normalizing
-   plots, and allows ROI-based analysis.
+The `plotter` module provides an interactive visualization interface for `DataCube` objects
+in the HSI Wizard package. Users can step through wavelength layers, define regions of
+interest (ROIs), compute mean spectra, normalize data, and save or remove ROI-based
+spectral plots.
 
-   :param dc: DataCube containing hyperspectral imaging (HSI) data.
-   :type dc: DataCube
+Interactive Controls
+====================
 
-   This function allows interactive visualization of hyperspectral imaging data, providing features such as:
+- **Left/Right arrow keys**: Step through wavelength layers.
+- **Mouse click on image panel**: Select a single-pixel ROI at the cursor location.
+- **Rectangle drag on image panel**: Define a custom ROI region.
+- **Click on spectrum panel**: Jump to the nearest wavelength layer based on clicked wavelength.
+- **Save Plot button**: Save the current ROI mean spectrum, overlaid with a unique semi-transparent color.
+- **Remove Plot button**: Remove the most recently saved ROI spectrum and its rectangle overlay.
+- **Normalize Y (0–1) checkbox**: Toggle normalization of all spectra on the Y-axis between 0 and 1.
 
-   - Image layer selection by wavelength.
-   - ROI (Region of Interest) selection and mean spectrum calculation.
-   - Saving and removing plotted spectra with unique colors.
-   - Normalization toggle for spectral data.
+Function: wizard.plotter(dc)
+============================
 
-   The interactive interface includes:
+.. autofunction:: wizard._exploration.plotter.plotter
 
-   - A main image display of the selected wavelength layer.
-   - A spectrum plot with the mean spectrum of the selected ROI.
-   - Buttons to save and remove plots.
-   - A checkbox to enable normalization.
 
-   Example Usage:
-   --------------
+Example Usage
+=============
 
-   The following example script demonstrates how to use the `plotter` function:
+The following example demonstrates how to use the `plotter` function:
 
-   .. literalinclude:: ../../../../examples/01_plot_DataCube/01_example.py
-      :language: python
-      :linenos:
+.. literalinclude:: ../../../../examples/01_plot_DataCube/01_example.py
+   :language: python
+   :linenos:
 
-   .. figure:: ../../../../resources/imgs/plotter_example.png
-      :align: center
-      :alt: Example of the plotter function in action.
+.. figure:: ../../../../resources/imgs/plotter_example.png
+   :align: center
+   :alt: Example of the plotter function in action.
 
-      Example visualization of hyperspectral data using the plotter function.
+   Interactive example of hyperspectral data visualization using the `plotter` function.
