### 2025-05-23 f187d91
docs/ update plotter docstring to pass stylecheck
