### 2025-05-23 16664a0
docs: update plotter documentation and docstring, added an Interactive Controls section (closes #17) (#23)
* docs: update plotter documentation and docstring; added an Interactive Controls section (closes #17)

* docs/ update plotter docstring to pass stylecheck
