### 2025-05-23 dc36d09
docs: update surface documentation and docstring; added paper reference ; addresses #21
