+++ docs/source/modules/exploration/surface.rst
@@ -1,33 +1,39 @@
 .. _surface:
 
 Surface Plotter
+===============
+
+Overview
+--------
+
+The Surface Plotter module offers an interactive 3D visualization interface for DataCube objects. By combining slicing, threshold-based opacity control, and real-time rotation, it enables in-depth exploration of spatial and spectral features in chemical imaging datasets. This tool is particularly useful for applications such as mapping molecule concentration gradients in engineering processes.
+
+Function: wizard.plot_surface(dc)
+---------------------------------
+
+.. autofunction:: wizard._exploration.surface.plot_surface
+
+Paper Reference
 ---------------
 
-.. module:: wizard._exploration.surface
-   :platform: Unix
-   :synopsis: Plotter function for hyperspectral imaging (HSI) data.
+The Surface Plotter was employed in the study `Three-Dimensional, Molecule-Sensitive Mapping of Structured Falling Liquid Films Using Raman Scanning <https://doi.org/10.1002/cite.202300048>`_. Below is the graphical abstract linked to the full paper:
 
-.. function:: plot_surface(dc)
+.. figure:: https://onlinelibrary.wiley.com/cms/asset/82d41b4e-e935-4bc0-af07-07623713e130/cite202300048-fig-0002-m.jpg
+   :target: https://doi.org/10.1002/cite.202300048
+   :align: center
+   :alt: Graphical abstract of "Three-Dimensional, Molecule-Sensitive Mapping of Structured Falling Liquid Films Using Raman Scanning"
 
-   This module provides functionalities for manipulating and visualizing hyperspectral imaging (HSI) data cubes. It includes utilities for slicing, cutting, and plotting 3D surfaces interactively using sliders.
+   Three-Dimensional, Molecule-Sensitive Mapping of Structured Falling Liquid Films Using Raman Scanning, by `Nachtmann et al. <https://doi.org/10.1002/cite.202300048>`_
 
-   :param dc: DataCube containing hyperspectral imaging (HSI) data.
-   :type dc: DataCube
 
-   This function allows interactive visualization of hyperspectral imaging data, enabling users to explore data slices and apply threshold-based cuts interactively.
+Example Usage
+-------------
 
-   Example Usage:
-   --------------
+.. literalinclude:: ../../../../examples/06_plot_surface/06_example.py
+   :language: python
+   :linenos:
 
-   The following example script demonstrates how to use the `plot_surface` function:
-
-   .. literalinclude:: ../../../../examples/06_plot_surface/06_example.py
-      :language: python
-      :linenos:
-
-   .. figure:: ../../../../resources/imgs/surface_example.png
-      :align: center
-      :alt: Example of the surface plotter function in action.
-
-      Example visualization of hyperspectral data using the surface plotting function.
+.. figure:: ../../../../resources/imgs/surface_example.png
+   :align: center
+   :alt: Example of the surface plotter function in action.
 
