### 2025-05-23 25b9428
docs: update surface documentation and docstring; added paper reference ; addresses #21 (#24)
