+++ wizard/_exploration/surface.py
@@ -25,15 +25,39 @@ from matplotlib.widgets import Slider
 
 def dc_cut_by_value(z: np.array, val: int, type: str) -> np.array:
     """
-    Cut a data cube slice by a defined value.
+    Threshold a 2D data slice by a normalized cut-off value.
 
-    This function normalizes the input array, applies a threshold cut-off, and
-    replaces values below the cut-off with the minimum value of the array.
+    This function normalizes the input array to its maximum value, applies a threshold
+    such that all values less than or equal to the cut-off are replaced with the array's minimum,
+    and preserves values above the threshold. Use this to mask out low-intensity regions
+    in a hyperspectral data slice.
 
-    :param z: 2D numpy array representing a data cube slice
-    :param val: Cut-off value for thresholding
-    :param type: Type of cut-off operation (currently unused)
-    :return: Modified 2D numpy array after applying the cut-off
+    Parameters
+    ----------
+    z : np.ndarray
+        2D array (spatial slice) extracted from a DataCube (shape: (x, y)).
+    val : float
+        Normalized threshold between 0.0 and 1.0. Values less than or equal to this threshold
+        are set to the minimum value of the normalized slice.
+    type : str, optional
+        Label for the type of cut applied; currently unused but reserved for future modes.
+
+    Returns
+    -------
+    np.ndarray
+        A copy of the input array after normalization and threshold cut.
+
+
+    Notes
+    -----
+    - The input array is deeply copied to avoid modifying original data.
+    - Thresholding is performed after normalizing by the array's maximum.
+
+    Examples
+    --------
+    >>> slice2D = np.array([[0, 5], [10, 15]])
+    >>> dc_cut_by_value(slice2D, 0.5)
+    array([[0. , 0. ], [0.66666667, 1. ]])
     """
     new_z = copy.deepcopy(z)
     new_z /= new_z.max()
@@ -43,14 +67,32 @@ def dc_cut_by_value(z: np.array, val: int, type: str) -> np.array:
 
 def get_z_surface(cube: np.array, v: int) -> np.array:
     """
-    Calculate the surface for plotting from a data cube.
+    Extract the positive-valued surface from a spectral slice of a data cube.
 
-    This function extracts a slice from the data cube at the specified
-    index and generates a 2D surface array for plotting.
+    This function selects the v-th spectral band from a 3D hyperspectral cube,
+    masks out non-positive values, and constructs a 2D surface array for plotting.
+    Use this to visualize spatial distribution at a given wavelength index.
 
-    :param cube: 3D numpy array representing the data cube
-    :param v: Index of the slice to extract
-    :return: 2D numpy array representing the surface for plotting
+    Parameters
+    ----------
+    cube : np.ndarray
+        3D data cube with shape (v, x, y), where v is the number of spectral bands.
+    v : int
+        Index of the spectral band to extract (0 <= v < cube.shape[0]).
+
+    Returns
+    -------
+    np.ndarray
+        2D array (shape: (x, y)) containing only the positive values from the slice;
+        non-positive entries are left at zero.
+
+
+    Examples
+    --------
+    >>> cube = np.zeros((3, 2, 2))
+    >>> cube[1] = [[1, -1], [0, 2]]
+    >>> get_z_surface(cube, 1)
+    array([[1., 0.], [0., 2.]])
     """
     z = np.zeros((cube.shape[1], cube.shape[2]))
     slice_v = cube[v, :, :]
@@ -61,13 +103,35 @@ def get_z_surface(cube: np.array, v: int) -> np.array:
 
 def plot_surface(dc: DataCube, index: int = 0):
     """
-    Plot a 3D surface from a DataCube slice with interactive sliders.
+    Create an interactive 3D surface plot from a DataCube slice.
 
[…]
