### 2025-05-23 BlueSpacePotato/hsi-wizard#21 @BlueSpacePotato
Thank you for the feedback.

I have adapted the documentation to hopefully provide more clarity.

Unfortunately I can't use real-world data because I don't have the rights to share the data I have. Alternatively, I have linked the study for which I implemented the function with an example image. I hope that works too?

https://hsi-wizard.readthedocs.io/modules/exploration/surface.html#paper-reference
