### 2025-05-24 BlueSpacePotato/hsi-wizard#21 @mh-skjelvareid
Thanks! The updated documentation gives a lot more context. As I understand it, the tool is used to browse different wavelengths and test different cutoff values for separating background from the signal / object of interest. The only thing that still is a bit confusing is the use of a surface for plotting. The figure caption in the paper says "3D surface plot of sodium sulfate flowing over a pyramid structure on a falling film, measured with Raman imaging. The coloring refers to the height of the structure and is for better visibility." So in this this case the pixel value corresponds to a spatial dimension (z-axis) and not light intensity?
