### 2025-05-25 1345410
fix: removed plotter from template (adresses #20)
