### 2025-05-25 10fecc5
fix: removed plotter from 03 and 04 example #20
