+++ examples/04_use_Templates/04_example.py
@@ -13,6 +13,3 @@ dc.execute_template('04_example.yml')
 
 print('Processed Datacube:')
 print(dc)
-
-# plot data
-wizard.plotter(dc)
