### 2025-05-25 6226a43
Fix/create template 20 (#25)
* fix: removed plotter from template (adresses #20)

* fix: removed plotter from 03 and 04 example #20
