+++ examples/04_use_Templates/04_example.py
@@ -9,7 +9,7 @@ print('DataCube:')
 print(dc)
 
 # load and execute the template
-dc.execute_template('03_example.yml')
+dc.execute_template('04_example.yml')
 
 print('Processed Datacube:')
 print(dc)
