### 2025-05-25 BlueSpacePotato/hsi-wizard#20 @BlueSpacePotato
Thanks for pointing this out.  I removed the plotter from the template examples.

> but using the interactive window to show this is a bit clunky.

Any other suggestions? Using pyplot with subplots?

>  Also, when I close the window, another error pops up: " File "C:\Users\mha114\Miniconda3\envs\hsi-wizard\Lib\site-packages\joblib\externals\loky\backend\resource_tracker.py", line 181, in del, AttributeError: 'super' object has no attribute 'del'".

Thank you for that info. Sadly, I cannot reproduce your problem as I do not have an win OS. I would like to fix the problem. Can you give me more information? Does it only occur in this example or do you have other problems with the plotter?
