### 2025-05-26 d89a340
docs/update readme (addresses #19, #16)
