### 2025-05-26 1a3e238
docs/update readme (addresses #19, #16) (#27)
