+++ README.md
@@ -7,32 +7,47 @@
 
 # HSI Wizard
 
-See Beyond the Visible: The Magic of Hyperspectral Imaging
+See Beyond the Visible: The Magic of Hyperspectral Imaging for Medical & Bioinformatics Applications
 
 <img src="./resources/imgs/hsi_wizard_logo.svg" alt="hsi_wizard_logo" style="width: 100%">
 
 
+
 ## Introduction
 
-Welcome to the `hsi-wizard` package! This Python package provides a straightforward environment for hyperspectral imaging (HSI) analysis, supporting everything from basic spectral analysis to advanced machine learning and AI methods. Whether you're working with raw sensor data or pre-processed datasets, `hsi-wizard` offers a suite of tools to simplify and enhance your analysis workflow.
+Welcome to the `hsi-wizard` package! Designed primarily for **medical hyperspectral imaging** and **bioinformatics workflows**, this Python package provides a straightforward environment for hyperspectral imaging (HSI) analysis, supporting everything from basic spectral analysis to advanced machine learning and AI methods. Whether you're working with raw sensor data or pre-processed datasets, `hsi-wizard` offers a suite of tools to simplify and enhance your analysis workflow.
 
 If you're new here, the best place to start is the [documentation](https://hsi-wizard.readthedocs.io), where you'll find detailed instructions on how to begin.
 
-
 ## Features
-- DataCube Class for managing and processing HSI data.
-- Spectral plotting and visualization.
-- Clustering and spectral analytics.
-- Tools for merging and processing HSI data.
-- Data loaders for various file formats (e.g., NRRD, Pickle, TDMS, and XLSX).
-- Decorators for method tracking, input validation, and execution time logging.
+- DataCube Class for managing and processing HSI data
+- Spectral plotting and visualization
+- Clustering and spectral analytics
+- Tools for merging and processing HSI data
+- Data loaders for various file formats (e.g., NRRD, Pickle, TDMS, and XLSX)
+- Decorators for method tracking, input validation, and execution time logging
 
-## Requirements
-- [Python](https://www.python.org) >3.10
+## Geospatial Limitations
+This package does not track or process geographical coordinates. If you require full remote-sensing or GIS integration, you may need complementary tools. hsi-wizard focuses on medical and lab-based spectral analysis, so spatial georeferencing is restricted.
+
+
+
+## Comparison with Existing Tools
+| **Attribute**        | **HSI-Wizard**                                             | **PySptools**                                                                                                 | **HyDe**                                                                                  | **Spectral Python (SPy)**                                                                | **ENVI**                                                         |
+| -------------------- | ---------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------- | ----------------------------------------------------------------------------------------- | ---------------------------------------------------------------------------------------- | ---------------------------------------------------------------- |
+| **Category**         | Medical Imaging                                            | Geospatial & Research Spectral Analysis                                                                       | Preprocessing & Denoising                                                                 | General-Purpose Data I/O & Basic Handling                                                | Commercial All-rounder                                           |
+| **Key Features**     | End-to-end medical pipeline,  DataCube,  Merge & analytics | Endmember extraction, Spectral unmixing,  Low-rank wavelet & DNN denoising,  Energy-efficient implementations | Read/display/manipulate HSI files, Bip/Bil/Bsq interleaves, Basic classification routines | Advanced spectral processing & classification, Full GUI & Python API, ArcGIS integration |                                                                  |
[…]
