### 2025-05-26 BlueSpacePotato/hsi-wizard#21 @BlueSpacePotato
The height and color of the plot refer to the intensity/counts of a selected wavelength \lambda.

I suspect there was a problem with the interpretation of 'refers to'.  As far as I know, the author wanted to say that the geometric figure can be recognized by the intensities.
