### 2025-05-26 BlueSpacePotato/hsi-wizard#16 @BlueSpacePotato
Thank you for your feedback!
I've updated the “Contributing” section and set up a Github discussion for better support.
