### 2025-05-26 #8207 @mh-skjelvareid
I see that @BlueSpacePotato is working steadily on closing the comments / issues that I opened - nice!. In my feedback above I created issues for some points (mainly concrete technical issues / clarifications) and left other points as only a comment. However, I see that using issues is a very nicely structured way to address the review points. If it's OK with both @BlueSpacePotato  and @csoneson, I can open issues for the remaining points as well. In some cases it's maybe enough to discuss the points without making any changes - but using issues seems like a good way to structure it, anyway. What do you think?
