### 2025-05-26 #8207 @BlueSpacePotato
That's fine for me. I try to structure my issues as well as possible, so I would be happy to overview all issues directly.  If you have any suggestions / problems or other points, I would be happy to receive input and discuss the issues.

thanks for the feedback and good advice @mh-skjelvareid
