### 2025-05-26 BlueSpacePotato/hsi-wizard#20 @mh-skjelvareid
> > but using the interactive window to show this is a bit clunky.
>
> Any other suggestions? Using pyplot with subplots?

Using matplotlib / pyplot with before/after plots seems like a simple "read-only" solution.

> > Also, when I close the window, another error pops up: " File "C:\Users\mha114\Miniconda3\envs\hsi-wizard\Lib\site-packages\joblib\externals\loky\backend\resource_tracker.py", line 181, in del, AttributeError: 'super' object has no attribute 'del'".
>
> Thank you for that info. Sadly, I cannot reproduce your problem as I do not have an win OS. I would like to fix the problem. Can you give me more information? Does it only occur in this example or do you have other problems with the plotter?

The error does not occur when I run the simple example at https://hsi-wizard.readthedocs.io/examples/index.html#plot-datacube

I've also tried opening two plotter windows without doing any other processing:

    import numpy as np
    import wizard

    dc1 = wizard.DataCube(np.random.rand(20, 640, 460))
    wizard.plotter(dc1)
    dc2 = wizard.DataCube(np.random.rand(20, 640, 460))
    wizard.plotter(dc2)

... without any errors. So far the only problem was with running the (old) create template script. I'm running Python 3.13 in a conda environment.
