### 2025-05-27 BlueSpacePotato/hsi-wizard#28 @mh-skjelvareid
Review README regarding machine learning and AI
This issue is part of the review of the paper submitted to the Journal of Open Source Software: https://github.com/openjournals/joss-reviews/issues/8207

The README file states that the HSI-Wizard "supports everything from basic spectral analysis to advanced machine learning and AI methods." It is unclear what is meant by machine learning and AI in this context. The package supports clustering (unsupervised learning) and manipulation of spectral data, but "machine learning" and "AI" does not seem appropriate here. Except if the word "supports" is meant as enabling subsequent machine learning and AI (outside the package) by pre-processing hyperspectral data?
