### 2025-05-27 BlueSpacePotato/hsi-wizard#29 @mh-skjelvareid
Consider adding support for additional hyperspectral file formats
This issue is part of the review of the paper submitted to the Journal of Open Source Software: https://github.com/openjournals/joss-reviews/issues/8207

The JOSS paper emphasizes that HSI-Wizard is meant as tool to unify a diverse set of hyperspectral datasets. If this is the case, I think the list of supported file formats should be extended to include some more of the most common file formats for hyperspectral data. My background is not in medical imaging, so I don't know the most common file formats in this field, but I that think at least support for [ENVI files](https://www.nv5geospatialsoftware.com/docs/enviheaderfiles.html) would be a natural addition (which is relatively easy via [Spectral Python](http://www.spectralpython.net/fileio.html#envi-headers)). I don't see this as an acceptance-blocker, as I think the authors are free to set the scope of their own software, but it would significantly expand the audience for the package.
