### 2025-05-27 BlueSpacePotato/hsi-wizard#30 @mh-skjelvareid
Compare HSI-Wizard with other similar software
This issue is part of the review of the paper submitted to the Journal of Open Source Software: https://github.com/openjournals/joss-reviews/issues/8207

The paper should include a description of how this software compares to other commonly-used packages in this research area. I think the paper would benefit from a comparison with software such as

- [Spectral Python](https://www.spectralpython.net/)
- [Spectronon](https://resonon.com/blog-spectronon-hyperspectral-imaging-software)
- [IDCubeLite](https://www.idcubes.com/idcubelite)
- [GLIMPS](https://www.rese-apps.com/software/glimps/index.html)

Note that most of these are not open source - the authors can take the opportunity to argue the advantages of HSI-Wizard being open source.
