### 2025-05-27 BlueSpacePotato/hsi-wizard#31 @mh-skjelvareid
Fix paper DOI formatting and add link to software (GitHub / PyPI)
This issue is part of the review of the paper submitted to the Journal of Open Source Software: https://github.com/openjournals/joss-reviews/issues/8207

Some of the references in [paper.bib](https://github.com/BlueSpacePotato/hsi-wizard/blob/joss/joss/paper.bib) are not validly formatted (see editorialbot output). The [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) also states that the paper should include "A list of key references *including a link to the software archive.*".
