### 2025-05-27 BlueSpacePotato/hsi-wizard#32 @mh-skjelvareid
Expand statement of need, preferrably with example
This issue is part of the review of the paper submitted to the Journal of Open Source Software: https://github.com/openjournals/joss-reviews/issues/8207

The paper describes data fusion and logging of data manipulations as two of the most important features of the software. I think this description can be expanded to give the readers a better understanding of why this is needed, and how the `DataCube` class helps solve these problems. A short example using real-world data would be instructive (in the paper and/or in the README file).
