### 2025-05-27 BlueSpacePotato/hsi-wizard#32 @BlueSpacePotato
I'm trying to work out the advantages better. But to what extent I am able to provide datasets or share real data, I am not sure. Publishing medical datasets and the results is not always easy.
