### 2025-05-28 a903828
doc/better example and documenation
