+++ wizard/_processing/cluster.py
@@ -14,8 +14,8 @@ This module initializes the cluster functions of the hsi-wizard package.
 Functions
 ---------
 
-.. autofunction:: quit_low_change_in_clusters
-.. autofunction:: discard_clusters
+.. autofunction:: isodata
+.. autofunction:: smooth_kmneas
 
 
 Credits
@@ -24,8 +24,8 @@ The Isodata code was inspired by:
 - Repository: pyRadar
 - Author/Organization: PyRadar
 - Original repository: https://github.com/PyRadar/pyradar/
-
 """
+
 import numpy as np
 from scipy.cluster.vq import vq
 from typing import Tuple
@@ -35,14 +35,25 @@ from sklearn.metrics import pairwise_distances
 from scipy.signal import convolve2d
 
 
-def quit_low_change_in_clusters(centers: np.ndarray, last_centers: np.ndarray, theta_o: float) -> bool:
+def _quit_low_change_in_clusters(centers: np.ndarray, last_centers: np.ndarray, theta_o: float) -> bool:
     """
-    Stop algorithm by low change in the clusters values between each iteration.
+    Determine if cluster update change is below the threshold to stop iteration.
 
-    :param centers: Cluster centers
-    :param last_centers: Last cluster centers
-    :param theta_o: threshold change in the clusters between each iter
-    :return: True if it should stop, otherwise False.
+    Compares current and previous cluster centers and stops if all changes are below `theta_o`.
+
+    Parameters
+    ----------
+    centers : np.ndarray
+        Current cluster centers.
+    last_centers : np.ndarray
+        Cluster centers from the previous iteration.
+    theta_o : float
+        Threshold for the percent change in cluster centers.
+
+    Returns
+    -------
+    bool
+        True if change is below the threshold and iteration should stop, False otherwise.
     """
     qt = False
     if centers.shape == last_centers.shape:
@@ -54,15 +65,25 @@ def quit_low_change_in_clusters(centers: np.ndarray, last_centers: np.ndarray, t
     return qt
 
 
-def discard_clusters(img_class_flat: np.ndarray, centers: np.ndarray, clusters_list: np.ndarray, theta_m: int) -> Tuple[np.ndarray, np.ndarray, int]:
+def _discard_clusters(img_class_flat: np.ndarray, centers: np.ndarray, clusters_list: np.ndarray, theta_m: int) -> Tuple[np.ndarray, np.ndarray, int]:
     """
-    Discard clusters with fewer than theta_m.
+    Remove clusters with fewer members than the minimum threshold.
 
-    :param img_class_flat: Classes of the flatten image
-    :param centers: Cluster centers
-    :param clusters_list: List of clusters
-    :param theta_m: threshold value for min number in each cluster
-    :return: Tuple of the new cluster centers, a list of the new clusters and a new value for k_
+    Parameters
+    ----------
+    img_class_flat : np.ndarray
+        Flattened image class labels.
+    centers : np.ndarray
+        Cluster centers.
+    clusters_list : np.ndarray
+        List of cluster labels.
+    theta_m : int
+        Minimum number of pixels per cluster.
+
+    Returns
+    -------
+    Tuple[np.ndarray, np.ndarray, int]
+        Updated centers, cluster list, and number of clusters.
     """
     k_ = centers.shape[0]
     to_delete = np.array([])
@@ -89,15 +110,25 @@ def discard_clusters(img_class_flat: np.ndarray, centers: np.ndarray, clusters_l
     return new_centers, new_clusters_list, k_
 
 
-def update_clusters(img_flat: np.ndarray, img_class_flat: np.ndarray, centers: np.ndarray, clusters_list: np.ndarray) -> Tuple[np.ndarray, np.ndarray, int]:
[…]
