+++ examples/04_use_Templates/04_example.yml
@@ -1,8 +1,0 @@
-- method: remove_spikes
-  kwargs:
-    threshold: 6500
-    window: 3
-- method: resize
-  kwargs:
-    x_new: 540
-    y_new: 360
