### 2025-05-28 de54bac
doc/better example and documenation (#33)
