+++ docs/requirements.txt
@@ -1,2 +1,3 @@
 sphinx-rtd-theme
+sphinx-exec-code
 hsi-wizard
