+++ docs/source/modules/exploration/surface.rst
@@ -36,4 +36,5 @@ Example Usage
 .. figure:: ../../../../resources/imgs/surface_example.png
    :align: center
    :alt: Example of the surface plotter function in action.
+   :caption: Example of the surface plotter function in action.
 
