### 2025-05-29 BlueSpacePotato/hsi-wizard#32 @mh-skjelvareid
If you can't use "your own" data, consider using an example from this dataset: https://www.heiporspectral.org/
