### 2025-05-29 BlueSpacePotato/hsi-wizard#28 @mh-skjelvareid
I fully agree that clustering and supervised learning is a subset of machine learning and AI. I guess my point is that the description gives the impression that the package includes a large set of ML/AI features. But it's not very important, keep it if you want. 🙂
