### 2025-06-03 6f25db2
feature/added envi/hdr file support, added code test and docs and a example
