+++ docs/source/modules/processing/cluster.rst
@@ -13,22 +13,8 @@ Module Overview
 Functions
 ---------
 
-.. autofunction:: quit_low_change_in_clusters
-.. autofunction:: discard_clusters
-.. autofunction:: update_clusters
-.. autofunction:: initial_clusters
-.. autofunction:: sort_arrays_by_first
-.. autofunction:: split_clusters
-.. autofunction:: compute_avg_distance
-.. autofunction:: compute_overall_distance
-.. autofunction:: merge_clusters
 .. autofunction:: compute_pairwise_distances
 .. autofunction:: isodata
 
-
-Credits
--------
-The Isodata code was inspired by:
- Repository: pyRadar
- Author/Organization: PyRadar
- Original repository: https://github.com/PyRadar/pyradar/
+.. note::
+  The Isodata code was inspired by `pyRadar <https://github.com/PyRadar/pyradar/>` from PyRadar.
