+++ pyproject.toml
@@ -43,6 +43,7 @@ dependencies = [
     "rembg",
     "onnxruntime",
     "scikit-learn",
+    "spectral",
 ]
 
 [project.optional-dependencies]
