### 2025-06-03 222839e
feature/added envi/hdr file support, added code test and docs and a example (#34)
