### 2025-06-03 BlueSpacePotato/hsi-wizard#29 @BlueSpacePotato
I have integrated the .envi file support via Spectral.

Due to the lack of a data set, I used a some data from [NV5](https://www.nv5geospatialsoftware.com/docs/HyperspectralAnalysisTutorial.html). I hope that fits.

![Image](https://github.com/user-attachments/assets/e28bf9ce-12fe-414b-b710-17aba468d7f4)
