+++ README.md
@@ -37,7 +37,7 @@ This package does not track or process geographical coordinates. If you require 
 | -------------------- | ---------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------- | ----------------------------------------------------------------------------------------- | ---------------------------------------------------------------------------------------- | ---------------------------------------------------------------- |
 | **Category**         | Medical Imaging                                            | Geospatial & Research Spectral Analysis                                                                       | Preprocessing & Denoising                                                                 | General-Purpose Data I/O & Basic Handling                                                | Commercial All-rounder                                           |
 | **Key Features**     | End-to-end medical pipeline,  DataCube,  Merge & analytics | Endmember extraction, Spectral unmixing,  Low-rank wavelet & DNN denoising,  Energy-efficient implementations | Read/display/manipulate HSI files, Bip/Bil/Bsq interleaves, Basic classification routines | Advanced spectral processing & classification, Full GUI & Python API, ArcGIS integration |                                                                  |
-| **File Support**     | Images, CSV, FSM, NRRD, Pickle (.pkl), TDMS, XLSX          | ENVI (.hdr/.img) via Spectral Python integration                                                              | None built-in (operates on NumPy arrays from user’s loader)                               | ENVI (.hdr/.img), BIL, BIP, BSQ interleaves                                              | ENVI (.hdr/.img), GeoTIFF, HDF, ASCII – broad commercial support |
+| **File Support**     | ENVI (.hdr/.img), Images, CSV, FSM, NRRD, Pickle (.pkl), TDMS, XLSX          | ENVI (.hdr/.img) via Spectral Python integration                                                              | None built-in (operates on NumPy arrays from user’s loader)                               | ENVI (.hdr/.img), BIL, BIP, BSQ interleaves                                              | ENVI (.hdr/.img), GeoTIFF, HDF, ASCII – broad commercial support |
 | **Licensing / Cost** | Open-source (MIT)                                          | Open-source (MIT)                                                                                             | Open-source (BSD-3-Clause)                                                                | Open-source (BSD-3)                                                                      | Proprietary, paid                                                |
 
 
