### 2025-06-03 542434b
doc/tryed to adress uses more cleary (#35)
