+++ docs/source/wizard/index.rst
@@ -1,31 +1,31 @@
 .. _wizard:
 
+Overview
 ========
-Features
-========
+
 The core functionality of `hsi-wizard` includes:
 
-- **DataCube Class**: Manage, track and process HSI data using a powerful object model
-- **Spectral Visualization**: Plot and explore spectral data using interactive and static visualizations
-- **Clustering and Analytics**: Apply clustering methods and analytics for exploratory data analysis
-- **File Format Support**: Seamless integration with various file formats, including `.nrrd`, `.pickle`, `.csv`, and `.xlsx`
+- **DataCube Class**: Manage, track, and process HSI data using a powerful object-oriented model
+- **Spectral Visualization**: Plot and explore spectral data through interactive and static visualizations
+- **Clustering and Analytics**: Apply clustering techniques and perform exploratory data analysis
+- **File Format Support**: Seamless integration with various formats, including `.nrrd`, `.pickle`, `.csv`, and `.xlsx`
 
 For a detailed breakdown of each module, refer to the :ref:`Modules <Modules>` section below.
 
-============
 Installation
 ============
-You can install `hsi-wizard` through pip or compile it from source:
 
-Via pip
--------
-To install the latest release via pip, use:
+You can install `hsi-wizard` via pip or compile it from source.
+
+Install via pip
+---------------
+To install the latest release using pip:
 
 .. code-block:: console
 
    pip install hsi-wizard
 
-Compile from Source
+Install from Source
 -------------------
 To install from source:
 
@@ -33,30 +33,97 @@ To install from source:
 
    python -m pip install -U pip setuptools wheel             # Install/update build tools
    git clone https://github.com/BlueSpacePotato/hsi-wizard   # Clone the repository
-   cd hsi-wizard                                             # Navigate into the directory
-   python -m venv .env                                       # Create a virtual environment
-   source .env/bin/activate                                  # Activate the environment
+   cd hsi-wizard                                             # Navigate into the project directory
+   python -m venv .venv                                      # Create a virtual environment
+   source .venv/bin/activate                                 # Activate the environment
    pip install -e .                                          # Install in editable mode
 
 
-====================
-DataCube Key Concept
-====================
+Comparison with with Existing Tools
+===================================
 
-DataCube Representation
------------------------
+.. list-table:: Comparison of Hyperspectral Imaging (HSI) Tools
+   :header-rows: 1
 
+   * - Attribute
+     - HSI-Wizard
+     - PySptools
+     - HyDe
+     - Spectral Python (SPy)
+     - ENVI
+   * - Category
+     - Medical Imaging
+     - Geospatial & Research
+       Spectral Analysis
+     - Preprocessing & Denoising
+     - General-Purpose Data
+       I/O & Basic Handling
+     - Commercial All-rounder
+   * - Key Features
+     - End-to-end medical pipeline,
+       DataCube support, merging & analytics
+     - Endmember extraction,
+       spectral unmixing,
+       wavelet & DNN denoising,
+       energy-efficient implementations
+     - Read/display/manipulate HSI files,
+       BIP/BIL/BSQ formats,
+       basic classification tools
+     - Advanced spectral
+       processing & classification,
+       full GUI & Python API,
+       ArcGIS integration
+     - ENVI (.hdr/.img), GeoTIFF, HDF,
+       ASCII – extensive commercial support
+   * - File Support
+     - ENVI (.hdr/.img), images, CSV, FSM,
+       NRRD, Pickle (.pkl), TDMS, XLSX
+     - ENVI (.hdr/.img)
+     - None built-in (requires user-defined NumPy loaders)
[…]
