### 2025-06-12 7430cef
fix/added comparison 30
