+++ joss/paper.bib
@@ -1,29 +1,39 @@
+@misc{your_package_name,
+  author = {Your Name},
+  title = {Software Title},
+  year = {2024},
+  publisher = {Python Package Index (PyPI)},
+  url = {https://pypi.org/project/your-package-name/},
+  note = {Version X.X.X, accessed: 2024-06-10}
+}
+
+
 @article{Bhargava,
-title = {Hyperspectral imaging and its applications: A review},
-journal = {Heliyon},
-volume = {10},
-number = {12},
-pages = {e33208},
-year = {2024},
-issn = {2405-8440},
-doi = {https://doi.org/10.1016/j.heliyon.2024.e33208},
-url = {https://www.sciencedirect.com/science/article/pii/S2405844024092399},
-author = {Anuja Bhargava and Ashish Sachdeva and Kulbhushan Sharma and Mohammed H. Alsharif and Peerapong Uthansakul and Monthippa Uthansakul},
-keywords = {Hyperspectral, Imaging, Spectral, Spatial, Temporal, Sensors}
+  title = {Hyperspectral imaging and its applications: A review},
+  journal = {Heliyon},
+  volume = {10},
+  number = {12},
+  pages = {e33208},
+  year = {2024},
+  issn = {2405-8440},
+  doi = {10.1016/j.heliyon.2024.e33208},
+  url = {https://www.sciencedirect.com/science/article/pii/S2405844024092399},
+  author = {Anuja Bhargava and Ashish Sachdeva and Kulbhushan Sharma and Mohammed H. Alsharif and Peerapong Uthansakul and Monthippa Uthansakul},
+  keywords = {Hyperspectral, Imaging, Spectral, Spatial, Temporal, Sensors}
 }
 
 @article{Guolan,
-author = {Guolan Lu and Baowei Fei},
-title = {{Medical hyperspectral imaging: a review}},
-volume = {19},
-journal = {Journal of Biomedical Optics},
-number = {1},
-publisher = {SPIE},
-pages = {010901},
-keywords = {Tissues, Tissue optics, Reflectivity, Hyperspectral imaging, Luminescence, Surgery, Charge-coupled devices, Tumors, Imaging systems, Cancer},
-year = {2014},
-doi = {10.1117/1.JBO.19.1.010901},
-URL = {https://doi.org/10.1117/1.JBO.19.1.010901}
+  author = {Guolan Lu and Baowei Fei},
+  title = {{Medical hyperspectral imaging: a review}},
+  volume = {19},
+  journal = {Journal of Biomedical Optics},
+  number = {1},
+  publisher = {SPIE},
+  pages = {010901},
+  keywords = {Tissues, Tissue optics, Reflectivity, Hyperspectral imaging, Luminescence, Surgery, Charge-coupled devices, Tumors, Imaging systems, Cancer},
+  year = {2014},
+  doi = {10.1117/1.JBO.19.1.010901},
+  URL = {https://doi.org/10.1117/1.JBO.19.1.010901}
 }
 
 @article{Burke,
@@ -41,56 +51,56 @@ URL = {https://doi.org/10.1117/1.JBO.19.1.010901}
 }
 
 @article{Knottnerus,
-title = {Promoting transparency of research and data needs much more attention},
-journal = {Journal of Clinical Epidemiology},
-volume = {70},
-pages = {1-3},
-year = {2016},
-issn = {0895-4356},
-doi = {https://doi.org/10.1016/j.jclinepi.2016.01.007},
-url = {https://www.sciencedirect.com/science/article/pii/S0895435616000159},
[…]
