### 2025-06-12 0d16274
added a uniform_filter method aswell as aditional cluster methods
