### 2025-06-13 837839c
feature/added smoothe cluster method
