### 2025-06-13 bf6d5be
feature/added normal kmeans
