### 2025-06-13 abdddab
Feature/new cluster methods (#36)
* added a uniform_filter method aswell as aditional cluster methods

* feature/added smoothe cluster method

* feature/added normal kmeans

* tests/added cluster tests

* tests/added additional datacube_ops tests
