### 2025-06-23 BlueSpacePotato/hsi-wizard#37 @tomelse
Issue installing
Re: https://github.com/openjournals/joss-reviews/issues/8207

I have tried installing `hsi-wizard` on my computer (Mac, M4 processor, Python 3.12.9), but come unstuck due to an error in settling the Numba dependency.

This is how I attempted to install the package:

` uv init . ` (Initialise a virtual environment in the current working directory).
` uv add hsi-wizard ` (Should be equivalent to pip install).

Unfortunately, it seems to be trying to install an older version of Numba, suggesting that some of the dependencies are not satisfied. Please could you look into this?

```
Using CPython 3.12.9
Creating virtual environment at: .venv
Resolved 65 packages in 558ms
  × Failed to build `numba==0.53.1`
  ├─▶ The build backend returned an error
  ╰─▶ Call to `setuptools.build_meta:__legacy__.build_wheel` failed (exit status: 1)

      [stderr]
      Traceback (most recent call last):
        File "<string>", line 14, in <module>
        File "/Users/telse/.cache/uv/builds-v0/.tmpCNtSID/lib/python3.12/site-packages/setuptools/build_meta.py", line 331, in get_requires_for_build_wheel
          return self._get_build_requires(config_settings, requirements=[])
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
        File "/Users/telse/.cache/uv/builds-v0/.tmpCNtSID/lib/python3.12/site-packages/setuptools/build_meta.py", line 301, in _get_build_requires
          self.run_setup()
        File "/Users/telse/.cache/uv/builds-v0/.tmpCNtSID/lib/python3.12/site-packages/setuptools/build_meta.py", line 512, in run_setup
          super().run_setup(setup_script=setup_script)
        File "/Users/telse/.cache/uv/builds-v0/.tmpCNtSID/lib/python3.12/site-packages/setuptools/build_meta.py", line 317, in run_setup
          exec(code, locals())
        File "<string>", line 50, in <module>
        File "<string>", line 47, in _guard_py_ver
      RuntimeError: Cannot install on Python version 3.12.9; only versions >=3.6,<3.10 are supported.

      hint: This usually indicates a problem with the package or the build environment.
  help: If you want to add the package regardless of the failed resolution, provide the `--frozen` flag to skip locking and syncing.
```
