### 2025-06-23 BlueSpacePotato/hsi-wizard#37 @tomelse
I see a similar issue on my Linux system (64 bit architecture, Python 3.12.3).

`uv init .`
`uv add hsi-wizard`

This seems to give a problem with llvmlite:

```Using CPython 3.12.3 interpreter at: /usr/bin/python3.12
Creating virtual environment at: .venv
Resolved 65 packages in 149ms
  × Failed to build `llvmlite==0.36.0`
  ├─▶ The build backend returned an error
  ╰─▶ Call to `setuptools.build_meta:__legacy__.build_wheel` failed (exit
      status: 1)

      [stderr]
      Traceback (most recent call last):
        File "<string>", line 14, in <module>
        File
      "/home/telse/.cache/uv/builds-v0/.tmpq5FprQ/lib/python3.12/site-packages/setuptools/build_meta.py",
      line 331, in get_requires_for_build_wheel
          return self._get_build_requires(config_settings, requirements=[])
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
        File
      "/home/telse/.cache/uv/builds-v0/.tmpq5FprQ/lib/python3.12/site-packages/setuptools/build_meta.py",
      line 301, in _get_build_requires
          self.run_setup()
        File
      "/home/telse/.cache/uv/builds-v0/.tmpq5FprQ/lib/python3.12/site-packages/setuptools/build_meta.py",
      line 512, in run_setup
          super().run_setup(setup_script=setup_script)
        File
      "/home/telse/.cache/uv/builds-v0/.tmpq5FprQ/lib/python3.12/site-packages/setuptools/build_meta.py",
      line 317, in run_setup
          exec(code, locals())
        File "<string>", line 55, in <module>
        File "<string>", line 52, in _guard_py_ver
      RuntimeError: Cannot install on Python version 3.12.3; only versions
      >=3.6,<3.10 are supported.

      hint: This usually indicates a problem with the package or the build
      environment.
  help: If you want to add the package regardless of the failed resolution,
        provide the `--frozen` flag to skip locking and syncing.
```
