### 2025-06-23 BlueSpacePotato/hsi-wizard#37 @tomelse
Same goes if I checkout the git repository and try to install locally `uv add ../hsi-wizard'.
