### 2025-06-23 #8207 @tomelse
Here is my feedback that does not reach the requirements of the checklist above:

# Functionality

* Installation: installation does not work for me as described (https://github.com/BlueSpacePotato/hsi-wizard/issues/37)
* Functionality: unable to confirm the functional claims since I can't install it.

# Documentation

* Statement of need: the documentation does not make it clear what problem the software is designed to solve. Please elaborate on what the alternatives are (e.g. directly using NumPy, I suppose?) and why hsi-wizard makes that easier?
