### 2025-06-23 BlueSpacePotato/hsi-wizard#38 @tomelse
Please add tests for other Python versions and operating systems
https://github.com/openjournals/joss-reviews/issues/8207

Please expand automated testing to run on Mac, Windows and other Python versions. If possible, please include both x86-64 and ARM (Mac M1) architectures.
