### 2025-06-23 0982fcd
adding overrides for uv and petry dependencies #37
