### 2025-06-23 5f4c473
adding overrides for uv and petry dependencies #37 (#39)
