### 2025-06-24 9130f5f
fix/updated build requirements
