### 2025-06-24 c1eaa4e
Fix/adding override for poetry and uv dependencies #37 (#40)
* adding overrides for uv and petry dependencies #37

* fix/updated build requirements
