+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools", "setuptools-scm"]
+requires = ["setuptools", "setuptools-scm", "numba>=0.56.4", "llvmlite >=0.44.0"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
@@ -54,7 +54,3 @@ build = ["twine", "build"]
 Homepage = "https://github.com/BlueSpacePotato/hsi-wizard"
 Issues = "https://github.com/BlueSpacePotato/hsi-wizard/issues"
 
-[tool.poetry.overrides]
-# Force sub-deps to latest wheel-compatible versions
-numba    = ">=0.56.4"
-llvmlite = ">=0.44.0"
