### 2025-06-24 6cf1ab4
fix/added numba to dependenccies
