+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools", "setuptools-scm", "numba>=0.56.4", "llvmlite >=0.44.0"]
+requires = ["setuptools", "setuptools-scm"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
@@ -27,6 +27,7 @@ classifiers = [
 
 dependencies = [
 # Your regular dependencies go here
+    "numba",
     "numpy",
     "matplotlib",
     "opencv-python",
