### 2025-06-24 07bf39f
Fix/adding override for poetry and uv dependencies #37 (#41)
* adding overrides for uv and petry dependencies #37

* fix/updated build requirements

* fix/added numba to dependenccies
