### 2025-06-24 BlueSpacePotato/hsi-wizard#37 @BlueSpacePotato
I’ve temporarily added Numba to my project’s dependencies so `uv add hsi-wizard` run correctly. I’m not thrilled about including a package I don’t actually use directly, but this workaround should keep everything working for now.
