### 2025-06-24 233adf3
test/added additional python and os versions to workflows
