+++ .github/workflows/main-branch.yml
@@ -8,7 +8,13 @@ on:
 jobs:
   test-and-coverage:
     name: Run tests and collect coverage
-    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest, windows-latest, macos-latest]
+        python-version: [3.10, 3.11, 3.12, 3.13]
+        arch: [x64, arm64]
+    runs-on: ${{ matrix.os }}
 
     steps:
       - name: Checkout
@@ -16,10 +22,14 @@ jobs:
         with:
           fetch-depth: 0
 
+      - name: Setup QEMU (for ARM on Ubuntu)
+        if: matrix.os == 'ubuntu-latest' && matrix.arch == 'arm64'
+        uses: docker/setup-qemu-action@v2
+
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.10"
+          python-version: ${{ matrix.python-version }}
 
       - name: Install dependencies
         run: pip install -e '.[tests]'
@@ -34,8 +44,8 @@ jobs:
 
   build-and-publish:
     name: Build and publish package
-    runs-on: ubuntu-latest
     needs: test-and-coverage
+    runs-on: ubuntu-latest
 
     steps:
       - name: Checkout Repository
@@ -44,10 +54,10 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.10"
+          python-version: "3.13"
 
       - name: Install dependencies
-        run: pip install -e '.[build]'  # Ensure build dependencies are installed
+        run: pip install -e '.[build]'
 
       - name: Set up Git user
         run: |
@@ -110,10 +120,9 @@ jobs:
           git tag -a "$new_version" -m "Version $new_version"
           git push origin "$new_version"
 
-
       - name: Build package
         run: |
-          python -m build 
+          python -m build
 
       - name: Publish package to PyPI
         env:
