### 2025-06-24 2556054
tests/removed abs path test to prevent posix str error
