### 2025-06-24 0f99d44
Test/adding additional os and python tests #30  (#42)
* test/added additional python and os versions to workflows

* test/fixed typo

* tests/removed abs path test to prevent posix str error
