+++ .github/workflows/non-main.yml
@@ -7,49 +7,64 @@ on:
 
 jobs:
   style-check:
-    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest, windows-latest, macos-latest]
+        python-version: ['3.10', '3.11', '3.12', '3.13']
+        arch: [x64, arm64]
+    runs-on: ${{ matrix.os }}
 
     steps:
       - name: Checkout Repository
         uses: actions/checkout@v4
 
+      - name: Setup QEMU (for ARM on Ubuntu)
+        if: matrix.os == 'ubuntu-latest' && matrix.arch == 'arm64'
+        uses: docker/setup-qemu-action@v2
+
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.13"
+          python-version: ${{ matrix.python-version }}
 
       - name: Install dev dependencies
         run: pip install flake8 pydocstyle
 
       - name: Run pydocstyle
-        run: |
-          pydocstyle wizard
+        run: pydocstyle wizard
 
       - name: Run flake8
-        run: |
-          flake8 wizard
+        run: flake8 wizard
 
   test-and-coverage:
-    runs-on: ubuntu-latest
     needs: style-check
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest, windows-latest, macos-latest]
+        python-version: ['3.10', '3.11', '3.12', '3.13']
+        arch: [x64, arm64]
+    runs-on: ${{ matrix.os }}
 
     steps:
       - name: Checkout Repository
         uses: actions/checkout@v4
 
+      - name: Setup QEMU (for ARM on Ubuntu)
+        if: matrix.os == 'ubuntu-latest' && matrix.arch == 'arm64'
+        uses: docker/setup-qemu-action@v2
+
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.10"
+          python-version: ${{ matrix.python-version }}
 
       - name: Install dependencies
         run: pip install -e '.[tests]'
 
       - name: Run tests
-        run: |
-          pytest tests/
+        run: pytest tests/
 
       - name: Generate coverage report
-        run: |
-          pytest --cov --cov-report=xml
-
+        run: pytest --cov --cov-report=xml
