### 2025-06-24 BlueSpacePotato/hsi-wizard#43 @tomelse
Questions about DataCube API.
I notice that you define an `__add__` function in your DataCube class for concatenating datasets. Given NumPy is a de facto standard for numerical Python libraries, do you think it is wise to diverge from the standard behaviour that one expects for adding arrays (i.e. element wise adding)? Is there a particular need for this shorthand in hyperspectral imaging datasets?

https://github.com/openjournals/joss-reviews/issues/8207
