### 2025-06-24 BlueSpacePotato/hsi-wizard#44 @tomelse
Plotting warning
I get the following warning when running the example below.

UserWarning: This figure includes Axes that are not compatible with tight_layout, so results might be incorrect.
  plt.tight_layout(rect=[0, 0, .95, 1])

```
import wizard
from wizard._utils.example import generate_pattern_stack

# generate randome data
data = generate_pattern_stack(20, 600, 400)

# creating some radnome data
dc = wizard.DataCube(data)

# plot data
wizard.plotter(dc)
```

https://github.com/openjournals/joss-reviews/issues/8207
