### 2025-06-24 BlueSpacePotato/hsi-wizard#43 @BlueSpacePotato
How do I deviate from the standard variarantee of NumPy? Sorry, maybe I'm so deep in my world that I don't understand the question properly.

The idea for `__add__` method is to put two DataCubes together. This includes the wavelengths and the actual cube data. In our current experiments, we are working a lot on adding value by merging complicated datasets, for which the function provides a good solution.

`__next__` and `__getitem__` should of course return the same. I'll take care of that.
