### 2025-06-24 BlueSpacePotato/hsi-wizard#43 @tomelse
I notice you dynamically add your data cube functions (background removal etc). I wonder if this is the best way to do it? Notably, this means that the functions don't appear in the API reference?

This is really just a suggestion - it doesn't necessarily need to change for the review, I just wondered why you made this design decision?
