### 2025-06-24 BlueSpacePotato/hsi-wizard#45 @tomelse
Some more examples would be great
There seem to be lots of interesting functions in `_processing/cluster.py`. I think it would be really nice to add a few more examples of using these to the documentation.

P.s. there's a typo in your kmeans function name `smooth_kmneas`.

https://github.com/openjournals/joss-reviews/issues/8207
