### 2025-06-24 de6c3c9
fix/update __next__ behavior (#46)
