### 2025-06-25 78457a5
fix/removed typo fro kmeans
