### 2025-06-25 d6c1d6e
feature/added n_iter loop to smoithing
