### 2025-06-26 470fbe6
doc/adding more cluster examples
