### 2025-06-26 42f8848
doc/addding more functions to cluster api
