+++ docs/source/modules/processing/cluster.rst
@@ -13,12 +13,21 @@ Module Overview
 Functions
 ---------
 
-.. autofunction:: compute_pairwise_distances
+.. autofunction:: kmeans
+.. autofunction:: smooth_kmeans
+
 .. autofunction:: isodata
-
 .. note::
-  The Isodata code was inspired by `pyRadar <https://github.com/PyRadar/pyradar/>` from PyRadar.
+  The Isodata code was inspired by `pyRadar <https://github.com/PyRadar/pyradar/>_` from PyRadar.
 
 
-  .. warning::
+
+.. autofunction:: spectral_spatial_kmeans
+
+.. autofunction:: spatial_agglomerative_clustering
+.. warning::
    Agglomerative clustering with spatial connectivity is conceptually elegant, but it doesn’t scale well to large 2D grids, and processing very large datasets can lead to high computing time.
+
+.. autofunction:: smooth_cluster
+
+.. autofunction:: pca
