### 2025-06-26 feed020
Doc/better examples 18 45 (#48)
* fix/removed typo fro kmeans

* feature/added n_iter loop to smoithing

* doc/adding more cluster examples

* doc/addding more functions to cluster api
