### 2025-06-26 BlueSpacePotato/hsi-wizard#45 @BlueSpacePotato
Thanks fot the input.

I fixed the typo and added few more examples.
