### 2025-07-08 #8207 @BlueSpacePotato
Thanks @csoneson for asking. Everything is fine and i hope that i will finish all the last issues this week.

Just two quick questions:

I dont understand this command:
> Yes, that's a common solution - pointing to either the GitHub repo or (e.g.) a PyPI landing page for the cited software in the bibliography.

And is there a resubmission process or will you just close this issue here and we are done?
