### 2025-07-10 8efb15f
fix/overlow error #50
