### 2025-07-15 37eb3f0
change default pollter color from wizard.plotter
