### 2025-07-15 c24bb9f
doc/added-heiprospectral-example
