### 2025-07-15 3a4b1b1
doc/added-heiprospectral-example (#52)
