+++ docs/source/examples/index.rst
@@ -52,4 +52,14 @@ Use Template
    templates/create_use
 
 
+Read World Examples
+===================
+
+.. toctree::
+   :maxdepth: 2
+   :titlesonly:
+
+   real_world/HeiPorSPECTRAL
+
+
 
