+++ joss/paper.md
@@ -22,50 +22,77 @@ bibliography: paper.bib
 
 # Summary
 
-`hsi-wizard` is an open-source Python package for processing, analyzing, and visualizing hyperspectral datasets. While its methods are adaptable to other fields, the core emphasis is on analytics for medical applications. Hyperspectral data, in this context, refers to images where each pixel contains multiple pieces of information, typically intensity values at different wavelengths. The number of spectral bands per pixel (n) can range from just a few to over 2000. These datasets originate from various sources, including scanners, imaging devices, and tomographic slides. Since hyperspectral data can come with different resolutions and pixel offsets, hsi-wizard enables seamless integration of multiple data sources. `hsi-wizard` provides an intuitive environment for hyperspectral analysis. With the `DataCube`, a standardized class for representing hyperspectral data is presented. This ensures uniform and consistent processing of hyperspectral data. In addition, `hsi-wizard` provides tools to manipulate, explore, and analyze `DataCubes`.  
+`hsi-wizard` is an open-source Python package designed for processing, analysing, and visualising hyperspectral datasets, primarily in medical applications. Hyperspectral data refers to images where each pixel contains multiple intensity values at different wavelengths, with the number of spectral bands ranging from only a few to well over a thousand. These datasets vary in resolution and structure, making analysis difficult due to inconsistent formats and limited reproducibility. To address these challenges, `hsi-wizard` provides a standardised data representation and analysis protocols that enhance reproducibility. The package features the `DataCube`, a standardised class for representing hyperspectral data, which simplifies data manipulation compared to traditional array-based tools like NumPy. The `DataCube` includes built-in methods tailored specifically for hyperspectral data, facilitating quicker exploration and reducing complexity for users.
 
-A key feature of the `DataCube` class is a protocol functionality that records the methods used from the `hsi-wizard` library. The protocol serves as a reusable template for other datasets, enabling researchers to reproduce results with minimal programming effort. Another feature of the `hsi-wizard` is a method for merging `DataCubes`. This makes it easy to combine different modalities into one large dataset. 
+A notable feature of the `DataCube` class is its protocol functionality, which records the methods used from the `hsi-wizard` library. This protocol acts as a reusable template for other datasets, allowing researchers, including those without programming experience, to replicate analyses accurately and effortlessly. Additionally, `hsi-wizard` allows users to merge `DataCubes` due to its standardised representation. This capability enables researchers, for example, to easily combine near-infrared (IR) and Raman spectroscopy datasets, resulting in more comprehensive and detailed tissue analyses. Such multimodal analysis simplifies the process and provides greater insights from combined data.
 
-The `hsi-wizard` simplifies the workflow for processing, analyzing, and visualizing hyperspectral data. It is designed to cater to both beginners and experts. Students can use the straightforward methods for educational purposes, while researchers can leverage advanced functions for professional studies, with the advantage that the results are reproducible and transparently documented.
+Overall, `hsi-wizard` streamlines the workflow for processing, analysing, and visualising hyperspectral data. It caters to both beginners and experts. Students can utilise straightforward methods for educational purposes, while researchers can leverage advanced functions for professional studies, ensuring that results are reproducible and transparently documented.
 
 # Statement of Need
 
[…]
