+++ joss/paper.bib
@@ -1,12 +1,10 @@
-@misc{your_package_name,
-  author = {Your Name},
-  title = {Software Title},
-  year = {2024},
-  publisher = {Python Package Index (PyPI)},
-  url = {https://pypi.org/project/your-package-name/},
-  note = {Version X.X.X, accessed: 2024-06-10}
+@misc{github_wizard,
+  author = {Wuehler, Felix},
+  title = {HSI-Wizard},
+  url = {https://github.com/BlueSpacePotato/hsi-wizard},
+  year = {2025},
+  note = {GitHub repository, Version 0.1.40},
 }
-
 
 @article{Bhargava,
   title = {Hyperspectral imaging and its applications: A review},
@@ -34,20 +32,6 @@
   year = {2014},
   doi = {10.1117/1.JBO.19.1.010901},
   URL = {https://doi.org/10.1117/1.JBO.19.1.010901}
-}
-
-@article{Burke,
-  title={Open science practices for eating disorders research},
-  author={Burke, Natasha L and Frank, Guido KW and Hilbert, Anja and Hildebrandt, Tom and Klump, Kelly L and Thomas, Jennifer J and Wade, Tracey D and Walsh, B Timothy and Wang, Shannon B and Weissman, Rachel S},
-  journal={International Journal of Eating Disorders},
-  volume={54},
-  number={10},
-  pages={1719--1729},
-  year={2021},
-  month={Oct},
-  doi={10.1002/eat.23607},
-  pmid={34555191},
-  pmcid={PMC9107337}
 }
 
 @article{Knottnerus,
@@ -125,4 +109,57 @@
   publisher  ={The Royal Society of Chemistry},
   doi  ={10.1039/D4AN00367E},
   url  ={http://dx.doi.org/10.1039/D4AN00367E},
+}
+
+@misc{SPy,
+  author       = {Boggs, E. and others},
+  title        = {Spectral Python (SPy): Python module for processing hyperspectral image data},
+  howpublished = {\url{https://www.spectralpython.net/}},
+  year         = {2025},
+  note         = {Open‑source MIT‑licensed package for hyperspectral I/O, visualization and classification}
+}
+
+@techreport{ENVI2018,
+  author       = {Wolfe, Jason D. and Black, Sarah R.},
+  title        = {Hyperspectral Analytics in ENVI: Target Detection and Spectral Mapping Methods},
+  institution  = {Harris Geospatial},
+  year         = {2018},
+  type         = {Technical Report},
+  note         = {Describes ENVI’s hyperspectral workflows including SAM, SFF, MNF etc.}
+}
+
+@article{Bakker2024,
+  author    = {Bakker, Wim and van Ruitenbeek, Frank and van der Werff, Harald and Hecker, Christoph and Dijkstra, Arjan and van der Meer, Freek},
+  title     = {Hyperspectral Python: HypPy},
+  journal   = {Algorithms},
+  volume    = {17},
+  number    = {8},
+  pages     = {337},
+  year      = {2024},
+  doi       = {10.3390/a17080337},
+  note      = {Open‑source Python package for hyperspectral image processing}
+}
+
+@misc{HyDe2022,
+  author       = {Coquelin, Daniel and Rasti, Behnood and G\"otz, Markus and Ghamisi, Pedram and Gloaguen, Richard and Streit, Achim},
+  title        = {HyDe: The First Open‑Source, Python‑Based, GPU‑Accelerated Hyperspectral Denoising Package},
+  howpublished = {ArXiv preprint},
+  year         = {2022},
+  note         = {GPU‑accelerated denoising toolbox with DNN and low‑rank methods}
+}
+
+
[…]
