### 2025-07-15 BlueSpacePotato/hsi-wizard#30 @BlueSpacePotato
Added a comparison section in the statement of need. 3a4b1b1
