### 2025-07-15 BlueSpacePotato/hsi-wizard#32 @BlueSpacePotato
Expanded the statement of need and added an example showcasing the heiprospectral dataset. 3a4b1b1
