### 2025-07-15 4ee8e13
paper/changed-kuemmel-refernece
