+++ joss/paper.bib
@@ -89,7 +89,7 @@
 
 @article{kummel,
   title={Rapid brain structure and tumour margin detection on whole frozen tissue sections by fast multiphotometric mid-infrared scanning},
-  author={Kümmel T. and van Marwick, Björn and Rittel, Miriam and Ramallo Guevara, Carina and Wühler, Fekux and Teumer, Tobias and Wängler, Björn and Hopf, Carsten and Rädle, Matthias},
+  author={Kuemmel T. and van Marwick, Björn and Rittel, Miriam and Ramallo Guevara, Carina and Wühler, Fekux and Teumer, Tobias and Wängler, Björn and Hopf, Carsten and Rädle, Matthias},
   journal={Scientific Reports},
   volume={11},
   number={1},
