### 2025-07-15 4ce34d6
paper/fix-error-kuemmel-refernce
