### 2025-07-17 BlueSpacePotato/hsi-wizard#30 @mh-skjelvareid
The documentation now includes [table comparing hsi-wizard with other software](https://hsi-wizard.readthedocs.io/wizard/index.html#comparison-with-with-existing-tools) and the paper has a paragraph which mentions other related software for processing hyperspectral data. Very good!

I think a paragraph should be added to the documentation explaining/summarizing how hsi-wizard relates to the other software. In what cases should you use hsi-wizard, and in what cases should you use something else? (it's good that the geospatial limitation is mentioned, by the way)

In the paper, some limitations are mentioned, but I think there should be a clearer description of what features are "missing" in other software and how this gap is filled with hsi-wizard.
