### 2025-07-17 BlueSpacePotato/hsi-wizard#55 @mh-skjelvareid
Format example code in paper to avoid long lines extending outside page
This issue is part of the review of the paper submitted to the Journal of Open Source Software: https://github.com/openjournals/joss-reviews/issues/8207

The example code in the paper has some lines that are so long that they extend outside the page. The code should be reformatted to stay within the right margin of the rest of the text.
