### 2025-07-17 BlueSpacePotato/hsi-wizard#56 @mh-skjelvareid
Demonstrate integration of multimodal datasets (image registration)
This issue is part of the review of the paper submitted to the Journal of Open Source Software: https://github.com/openjournals/joss-reviews/issues/8207

One of the main arguments in the statement of need in the paper is "hsi-wizard streamlines data handling and integrates multimodal datasets seamlessly". The integration of multimodal datasets probably refers to the various methods for image registration described under [datacube operations](https://hsi-wizard.readthedocs.io/modules/core/datacube_ops.html). However, there are no examples demonstrating this in the documentation or the paper.

The examples section should be expanded to include at least one example of integrating multimodal datasets - preferably with real data rather than synthetic (if possible).

You should also consider if this can be demonstrated in the paper.
