### 2025-07-21 f5e24a0
doc: added example print output
