### 2025-07-21 BlueSpacePotato/hsi-wizard#15 @BlueSpacePotato
Thanks for the feedback, good point. I added the print output to the examples.
