### 2025-07-21 BlueSpacePotato/hsi-wizard#56 @BlueSpacePotato
> One of the main arguments in the statement of need in the paper is "hsi-wizard streamlines data handling and integrates multimodal datasets seamlessly". The integration of multimodal datasets probably refers to the various methods for image registration described under [datacube operations](https://hsi-wizard.readthedocs.io/modules/core/datacube_ops.html).

If you look at the bigger picture, you are right. If you take just my phrasing, the `__add__` method from the `DataCube` would do the job too. If you find it helpfull i could provide an example merging two synthetic DataCubes.

> The examples section should be expanded to include at least one example of integrating multimodal datasets - preferably with real data rather than synthetic (if possible).

I spent quite some time searching for publicly available datasets that fit my needs. But I couldn't find any with complementary data. I work on another publication with such datasets, but I am not allowed to share the data until it is published. If this is done, I will greatly link my findings to an example.
