### 2025-07-21 b43db34
joss: shorten code lines
