### 2025-07-22 BlueSpacePotato/hsi-wizard#56 @mh-skjelvareid
Including an example in the docs would clarify exactly what is meant by integration of multimodal datasets, I think. Doing so would be very useful.

I think such an example could be "made" based on the same hyperspectral image from the HeiPorSpectral dataset that you've already used.

For example: You could extract some wavelength range from the image (e.g. 400-600 nm) and use it as one base "mode". Then, extract another range (e.g. 600-800 nm) and apply a number of augmentations (e.g. changing spatial resolution, translating and/or rotating the image, any "difference" that's realistic for modelling another mode). Then, finally, "add" the base image and the augmented image, demonstrating image registration in the process.

I think this would be more realistic than using synthetic data with squares, circles and triangles.

Make sense?
