### 2025-07-22 BlueSpacePotato/hsi-wizard#56 @BlueSpacePotato
I get your point, but I think this is still far from reality and not better than synthetic data.

In my experience, using different scanning methods results in several problems like different image ratios, background noises, switched contrasts (for example, when using reflection and transmission scanners), angles, as well as other effects like chromatic aberrations or different artefacts.  Taking just different subsets of wavelengths from the same dataset and using augmentation wouldn’t mirror these tasks.

I got a real-world example, and I linked publications for more usage references, which should be enough to fulfil the review criteria.

> Your paper must not focus on new research results accomplished with the software.

Therefore, to be honest, I would rather publish my package and, in the next step, show my results for the workflow and solutions in another publication.
