### 2025-07-22 BlueSpacePotato/hsi-wizard#56 @mh-skjelvareid
OK, I see your point, fair enough. If this feature is used in one or more of the publications you reference, maybe you could point that out explicitly in the paper?

I still think including some example that shows the "nuts and bolts" of this feature/these features would make sense (with synthesis data). But to be clear - these are just my two cents / suggestions, not a "showstopper".
