### 2025-07-24 BlueSpacePotato/hsi-wizard#30 @BlueSpacePotato
Any thoughts on this?
