### 2025-07-24 BlueSpacePotato/hsi-wizard#30 @mh-skjelvareid
![Screenshot_20250724-152821.png](https://github.com/user-attachments/assets/f117b615-c6df-4c64-ae63-6fa252d83a4b)

The documentation only has a table referencing other software, but no text, as far as I can see (see image)...?

If the paper text is now limited by word length it's probably fine as it is.
