### 2025-07-24 e969186
doc/added-merge-example
