### 2025-07-25 d72b56b
doc: added comparison description
