### 2025-07-25 278ad8f
Doc/add add example (#58)
* doc/added-merge-example (fix for #56)

* doc: added comparison description (closes #30)
