### 2025-07-25 BlueSpacePotato/hsi-wizard#30 @BlueSpacePotato
Thanks for your feedback. I added a few lines, hope this helps to clarify my tool for users.
