### 2025-07-25 4514195
joss: added following paper
