### 2025-07-25 BlueSpacePotato/hsi-wizard#55 @BlueSpacePotato
Thanks for the Input. Forgot to check for this.
b43db34 should solve the issue.
