### 2025-07-25 BlueSpacePotato/hsi-wizard#56 @BlueSpacePotato
Perfect, Thanks for the advice. I hope 4514195 and 278ad8f should fixed this topic.
