### 2025-07-30 #8207 @mh-skjelvareid
All my points have been adressed, and I've checked all the boxes in my checklist. I have nothing more to add. Good work!
