### 2025-08-11 BlueSpacePotato/hsi-wizard#60 @tomelse
Add NumPy import to the example script in the paper.
https://github.com/BlueSpacePotato/hsi-wizard/blob/a5f15ed83467943eea1c8a4c6cf2d4ad590315f6/joss/paper.md?plain=1#L58

I recommend adding `import numpy as np` to make sure that the code runs as listed in the paper.

Part of JOSS review: https://github.com/openjournals/joss-reviews/issues/8207
