### 2025-08-11 db12982
paper/added-np-import
