### 2025-08-11 BlueSpacePotato/hsi-wizard#60 @BlueSpacePotato
Thanks for the FB.

Fix it with db12982 and added the import.
