### 2025-08-17 #8207 @csoneson
👋🏻 @BlueSpacePotato - I have gone through your submission, and I have a few suggestions for the paper:

- line 41: should it be "The hsi-wizard package" (or just "hsi-wizard")?
- the section on lines 52-58 overlaps with the text on lines 41-46. Can lines 52-58 be removed?
- also the logging (lines 59-62) is already mentioned on line 46.
- line 73, and Figure 1 caption: hsi-wizard -> `hsi-wizard`
- Figure 1 caption: rois -> ROIs
- Figure 2 caption: You can include the Studier-Fischer et al reference without the `[]` to make it "Studier-Fischer et al (2023)"
- line 90: Github -> GitHub
- bibliography: put {} around surnames like 'van Marwick' to keep these together
- bibliography: letters with umlauts are sometimes written ü/ä, sometimes ue/ae - I'd suggest to make this consistent
- bibliography: for Manser et al, Nachtmann et al: case-protect to make raman -> Raman
- bibliography: same for Studier-Fischer et al: heidelberg -> Heidelberg
- the manuscript and GitHub repository use hsi-wizard, HSI Wizard, and HSI-wizard to refer to the software. Is this intentional, and how would you like users to refer to it?

Once these are addressed, you can generate a new proof with `@editorialbot generate pdf`. I will also make a separate post below with the other remaining steps.
