### 2025-08-19 aaaf4c8
paper/further-improvements
