+++ joss/paper.bib
@@ -47,7 +47,7 @@
 }
 
 @Article{heintz,
-  AUTHOR = {Heintz, Annabell and Sold, Sebastian and Wühler, Felix and Dyckow, Julia and Schirmer, Lucas and Beuermann, Thomas and Rädle, Matthias},
+  AUTHOR = {Heintz, Annabell and Sold, Sebastian and Wuehler, Felix and Dyckow, Julia and Schirmer, Lucas and Beuermann, Thomas and Raedle, Matthias},
   TITLE = {Design of a Multimodal Imaging System and Its First Application to Distinguish Grey and White Matter of Brain Tissue. A Proof-of-Concept-Study},
   JOURNAL = {Applied Sciences},
   VOLUME = {11},
@@ -60,8 +60,8 @@
 }
 
 @article{nachtmann,
-  author = {Nachtmann, Marcel and Feger, Daniel and Wühler, Felix and Scholl, Stephan and Rädle, Matthias},
-  title = {Three-Dimensional, Molecule-Sensitive Mapping of Structured Falling Liquid Films Using Raman Scanning},
+  author = {Nachtmann, Marcel and Feger, Daniel and Wuehler, Felix and Scholl, Stephan and Raedle, Matthias},
+  title = {Three-Dimensional, Molecule-Sensitive Mapping of Structured Falling Liquid Films Using {Raman} Scanning},
   journal = {Chemie Ingenieur Technik},
   volume = {96},
   number = {3},
@@ -75,7 +75,7 @@
 
 @Article{manser,
   AUTHOR = {Manser, Steffen and Keck, Shaun and Vitacolonna, Mario and Wuehler, Felix and Rudolf, Ruediger and Raedle, Matthias},
-  TITLE = {Innovative Imaging Techniques: A Conceptual Exploration of Multi-Modal Raman Light Sheet Microscopy},
+  TITLE = {Innovative Imaging Techniques: A Conceptual Exploration of Multi-Modal {Raman} Light Sheet Microscopy},
   JOURNAL = {Micromachines},
   VOLUME = {14},
   YEAR = {2023},
@@ -89,7 +89,7 @@
 
 @article{kummel,
   title={Rapid brain structure and tumour margin detection on whole frozen tissue sections by fast multiphotometric mid-infrared scanning},
-  author={Kuemmel, Tim and van Marwick, Björn and Rittel, Miriam and Ramallo Guevara, Carina and Wühler, Fekux and Teumer, Tobias and Wängler, Björn and Hopf, Carsten and Rädle, Matthias},
+  author={Kuemmel, Tim and {van Marwick}, Bjoern and Rittel, Miriam and Ramallo Guevara, Carina and Wuehler, Fekux and Teumer, Tobias and Waengler, Bjoern and Hopf, Carsten and Raedle, Matthias},
   journal={Scientific Reports},
   volume={11},
   number={1},
@@ -101,7 +101,7 @@
 
 
 @Article{vanmarwick,
-  author ={van Marwick, Björn and Kümmel, Tim and Wühler, Felix and Lauer, Felix and Hoffmann, Jan and Rädle, Matthias},
+  author ={{van Marwick}, Bjoern and Kuemmel, Tim and Wuehler, Felix and Lauer, Felix and Hoffmann, Jan and Raedle, Matthias},
   title  ={Rapid chemical detection and segmentation of latent fingerprints by means of a novel middle-infrared scanning method},
   journal  ={Analyst},
   year  ={2024},
@@ -150,8 +150,8 @@
 
 
 @article{Studier-Fischer2023,
[…]
