### 2025-08-27 #8207 @BlueSpacePotato
Hope i got everything correct.

### Additional Author Tasks After Review is Complete
---

> Double check authors and affiliations (including ORCIDs)
- [x] done

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
- [x] v0.1.42 [link](https://github.com/BlueSpacePotato/hsi-wizard/releases/tag/v0.1.42)

> Archive the release on Zenodo/figshare/etc and post the DOI here.
- [x] https://doi.org/10.5281/zenodo.16959455

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
- [x] done

> Make sure that the license listed for the archive is the same as the software license.
- [x] done

### Editor Tasks Prior to Acceptance
---

- [ ]  Set archive DOI with @editorialbot set <DOI here> as archive
- [ ] Set version with @editorialbot set <version here> as version
- [ ] Double check rendering of paper with @editorialbot generate pdf
- [ ] Specifically check the references with @editorialbot check references and ask author(s) to update as needed
- [ ] Recommend acceptance with @editorialbot recommend-accept
