### 2025-08-29 #8207 @csoneson
Thanks @BlueSpacePotato - a few minor things remaining:

- please update the title and author list of the Zenodo archive (in the record metadata) so that they agree with those of the paper
- I noticed that there is a bit of a mix between American and British spelling in the manuscript - could you make that consistent?
- it looks like the `van Marwick` et al reference has issue and page information available - could you add that to the bibliography?
- on line 83, I think there is one more instance where hsi-wizard should be `hsi-wizard`
