+++ joss/paper.bib
@@ -105,10 +105,13 @@
   title  ={Rapid chemical detection and segmentation of latent fingerprints by means of a novel middle-infrared scanning method},
   journal  ={Analyst},
   year  ={2024},
-  pages  ={-},
+  volume  ={149},
+  issue  ={24},
+  pages  ={5768-5783},
   publisher  ={The Royal Society of Chemistry},
   doi  ={10.1039/D4AN00367E},
   url  ={http://dx.doi.org/10.1039/D4AN00367E},
+
 }
 
 @misc{SPy,
