### 2025-08-31 #8207 @csoneson
@BlueSpacePotato Almost there! Could you just change the authors of the Zenodo record (`BlueSpacePotato` in the screenshot below) to match those of the papers too.
