### 2025-09-01 #8207 @BlueSpacePotato
> [@BlueSpacePotato](https://github.com/BlueSpacePotato) Almost there! Could you just change the authors of the Zenodo record (`BlueSpacePotato` in the screenshot below) to match those of the papers too.

Oh sorry, should be done @csoneson.
